@@ -1,21 +1,29 @@
-import React, { PropTypes } from 'react';
-import InlineSVG from 'react-inlinesvg';
-const exitUrl = require('../../../images/exit.svg');
+import PropTypes from 'prop-types';
+import React from 'react';
 import { Link } from 'react-router';
+import { withTranslation } from 'react-i18next';
 
 class ErrorModal extends React.Component {
-  componentDidMount() {
-    this.refs.modal.focus();
-  }
-
-
   forceAuthentication() {
     return (
       <p>
-        In order to save sketches, you must be logged in. Please&nbsp;
-        <Link to="/login" onClick={this.props.closeModal}>Login</Link>
-        &nbsp;or&nbsp;
-        <Link to="/signup" onClick={this.props.closeModal}>Sign Up</Link>.
+        {this.props.t('ErrorModal.MessageLogin')}
+        <Link to="/login" onClick={this.props.closeModal}> {this.props.t('ErrorModal.Login')}</Link>
+        {this.props.t('ErrorModal.LoginOr')}
+        <Link to="/signup" onClick={this.props.closeModal}>{this.props.t('ErrorModal.SignUp')}</Link>.
+      </p>
+    );
+  }
+
+  oauthError() {
+    const { t, service } = this.props;
+    const serviceLabels = {
+      github: 'GitHub',
+      google: 'Google'
+    };
+    return (
+      <p>
+        {t('ErrorModal.LinkMessage', { serviceauth: serviceLabels[service] })}
       </p>
     );
   }
@@ -23,8 +31,8 @@ class ErrorModal extends React.Component {
   staleSession() {
     return (
       <p>
-        It looks like you've been logged out. Please&nbsp;
-        <Link to="/login" onClick={this.props.closeModal}>log in</Link>.
+        {this.props.t('ErrorModal.MessageLoggedOut')}
+        <Link to="/login" onClick={this.props.closeModal}>{this.props.t('ErrorModal.LogIn')}</Link>.
       </p>
     );
   }
@@ -32,39 +40,39 @@ class ErrorModal extends React.Component {
   staleProject() {
     return (
       <p>
-        The project you have attempted to save is out of date. Please refresh the page.
+        {this.props.t('ErrorModal.SavedDifferentWindow')}
       </p>
     );
   }
 
   render() {
     return (
-      <section className="error-modal" ref="modal" tabIndex="0">
-        <header className="error-modal__header">
-          <h2 className="error-modal__title">Error</h2>
-          <button className="error-modal__exit-button" onClick={this.props.closeModal}>
-            <InlineSVG src={exitUrl} alt="Close Error Modal" />
-          </button>
-        </header>
-        <div className="error-modal__content">
-          {(() => { // eslint-disable-line
-            if (this.props.type === 'forceAuthentication') {
-              return this.forceAuthentication();
-            } else if (this.props.type === 'staleSession') {
-              return this.staleSession();
-            } else if (this.props.type === 'staleProject') {
-              return this.staleProject();
-            }
-          })()}
-        </div>
-      </section>
+      <div className="error-modal__content">
+        {(() => { // eslint-disable-line
+          if (this.props.type === 'forceAuthentication') {
+            return this.forceAuthentication();
+          } else if (this.props.type === 'staleSession') {
+            return this.staleSession();
+          } else if (this.props.type === 'staleProject') {
+            return this.staleProject();
+          } else if (this.props.type === 'oauthError') {
+            return this.oauthError();
+          }
+        })()}
+      </div>
     );
   }
 }
 
 ErrorModal.propTypes = {
-  type: PropTypes.string,
-  closeModal: PropTypes.func.isRequired
+  type: PropTypes.string.isRequired,
+  closeModal: PropTypes.func.isRequired,
+  t: PropTypes.func.isRequired,
+  service: PropTypes.string
 };
 
-export default ErrorModal;
+ErrorModal.defaultProps = {
+  service: ''
+};
+
+export default withTranslation()(ErrorModal);
