@@ -1,81 +1,92 @@
-import React, { PropTypes } from 'react';
-import InlineSVG from 'react-inlinesvg';
-const exitUrl = require('../../../images/exit.svg');
+import React from 'react';
+import { useTranslation } from 'react-i18next';
+import { metaKeyName, } from '../../../utils/metaKey';
 
-class KeyboardShortcutModal extends React.Component {
-  componentDidMount() {
-    this.isMac = navigator.userAgent.toLowerCase().indexOf('mac') !== -1;
-  }
-
-  render() {
-    return (
-      <section className="keyboard-shortcuts">
-        <header className="keyboard-shortcuts__header">
-          <h2>Keyboard Shortcuts</h2>
-          <button className="keyboard-shortcuts__close" onClick={this.props.closeModal}>
-            <InlineSVG src={exitUrl} alt="Close Keyboard Shortcuts Overlay" />
-          </button>
-        </header>
-        <ul title="keyboard shortcuts">
-          <li className="keyboard-shortcut-item">
-            <span className="keyboard-shortcut__command">Shift + Tab</span>
-            <span>Tidy</span>
-          </li>
-          <li className="keyboard-shortcut-item">
-            <span className="keyboard-shortcut__command">
-              {this.isMac ? 'Command + S' : 'Control + S'}
-            </span>
-            <span>Save</span>
-          </li>
-          <li className="keyboard-shortcut-item">
-            <span className="keyboard-shortcut__command">
-              {this.isMac ? 'Command + [' : 'Control + ['}
-            </span>
-            <span>Indent Code Left</span>
-          </li>
-          <li className="keyboard-shortcut-item">
-            <span className="keyboard-shortcut__command">
-              {this.isMac ? 'Command + ]' : 'Control + ]'}
-            </span>
-            <span>Indent Code Right</span>
-          </li>
-          <li className="keyboard-shortcut-item">
-            <span className="keyboard-shortcut__command">
-              {this.isMac ? 'Command + /' : 'Control + /'}
-            </span>
-            <span>Comment Line</span>
-          </li>
-          <li className="keyboard-shortcut-item">
-            <span className="keyboard-shortcut__command">
-              {this.isMac ? 'Command + Enter' : 'Control + Enter'}</span>
-            <span>Start Sketch</span>
-          </li>
-          <li className="keyboard-shortcut-item">
-            <span className="keyboard-shortcut__command">
-              {this.isMac ? 'Command + Shift + Enter' : 'Control + Shift + Enter'}
-            </span>
-            <span>Stop Sketch</span>
-          </li>
-          <li className="keyboard-shortcut-item">
-            <span className="keyboard-shortcut__command">
-              {this.isMac ? 'Command + Shift + 1' : 'Control + Shift + 1'}
-            </span>
-            <span>Toggle Text-based Canvas</span>
-          </li>
-          <li className="keyboard-shortcut-item">
-            <span className="keyboard-shortcut__command">
-              {this.isMac ? 'Command + Shift + 2' : 'Control + Shift + 2'}
-            </span>
-            <span>Turn Off Text-based Canvas</span>
-          </li>
-        </ul>
-      </section>
-    );
-  }
+function KeyboardShortcutModal() {
+  const { t } = useTranslation();
+  return (
+    <div className="keyboard-shortcuts">
+      <h3 className="keyboard-shortcuts__title">{t('KeyboardShortcuts.CodeEditing.CodeEditing')}</h3>
+      <p className="keyboard-shortcuts__description">
+        {t('KeyboardShortcuts.ShortcutsFollow')} <a href="https://shortcuts.design/toolspage-sublimetext.html" target="_blank" rel="noopener noreferrer">{t('KeyboardShortcuts.SublimeText')}</a>.
+      </p>
+      <ul className="keyboard-shortcuts__list">
+        <li className="keyboard-shortcut-item">
+          <span className="keyboard-shortcut__command">{'\u21E7'} + Tab</span>
+          <span>{t('KeyboardShortcuts.CodeEditing.Tidy')}</span>
+        </li>
+        <li className="keyboard-shortcut-item">
+          <span className="keyboard-shortcut__command">
+            {metaKeyName} + F
+          </span>
+          <span>{t('KeyboardShortcuts.CodeEditing.FindText')}</span>
+        </li>
+        <li className="keyboard-shortcut-item">
+          <span className="keyboard-shortcut__command">
+            {metaKeyName} + G
+          </span>
+          <span>{t('KeyboardShortcuts.CodeEditing.FindNextTextMatch')}</span>
+        </li>
+        <li className="keyboard-shortcut-item">
+          <span className="keyboard-shortcut__command">
+            {metaKeyName} + {'\u21E7'} + G
+          </span>
+          <span>{t('KeyboardShortcuts.CodeEditing.FindPreviousTextMatch')}</span>
+        </li>
+        <li className="keyboard-shortcut-item">
+          <span className="keyboard-shortcut__command">
+            {metaKeyName} + [
+          </span>
+          <span>{t('KeyboardShortcuts.CodeEditing.IndentCodeLeft')}</span>
+        </li>
+        <li className="keyboard-shortcut-item">
+          <span className="keyboard-shortcut__command">
+            {metaKeyName} + ]
+          </span>
+          <span>{t('KeyboardShortcuts.CodeEditing.IndentCodeRight')}</span>
+        </li>
+        <li className="keyboard-shortcut-item">
+          <span className="keyboard-shortcut__command">
+            {metaKeyName} + /
+          </span>
+          <span>{t('KeyboardShortcuts.CodeEditing.CommentLine')}</span>
+        </li>
+      </ul>
+      <h3 className="keyboard-shortcuts__title">General</h3>
+      <ul className="keyboard-shortcuts__list">
+        <li className="keyboard-shortcut-item">
+          <span className="keyboard-shortcut__command">
+            {metaKeyName} + S
+          </span>
+          <span>{t('Common.Save')}</span>
+        </li>
+        <li className="keyboard-shortcut-item">
+          <span className="keyboard-shortcut__command">
+            {metaKeyName} + Enter
+          </span>
+          <span>{t('KeyboardShortcuts.General.StartSketch')}</span>
+        </li>
+        <li className="keyboard-shortcut-item">
+          <span className="keyboard-shortcut__command">
+            {metaKeyName} + {'\u21E7'} + Enter
+          </span>
+          <span>{t('KeyboardShortcuts.General.StopSketch')}</span>
+        </li>
+        <li className="keyboard-shortcut-item">
+          <span className="keyboard-shortcut__command">
+            {metaKeyName} + {'\u21E7'} + 1
+          </span>
+          <span>{t('KeyboardShortcuts.General.TurnOnAccessibleOutput')}</span>
+        </li>
+        <li className="keyboard-shortcut-item">
+          <span className="keyboard-shortcut__command">
+            {metaKeyName} + {'\u21E7'} + 2
+          </span>
+          <span>{t('KeyboardShortcuts.General.TurnOffAccessibleOutput')}</span>
+        </li>
+      </ul>
+    </div>
+  );
 }
-
-KeyboardShortcutModal.propTypes = {
-  closeModal: PropTypes.func.isRequired
-};
 
 export default KeyboardShortcutModal;
