@@ -1,16 +1,36 @@
-import React, { PropTypes } from 'react';
+import PropTypes from 'prop-types';
+import React from 'react';
 import classNames from 'classnames';
-import InlineSVG from 'react-inlinesvg';
-// import SidebarItem from './SidebarItem';
-const folderUrl = require('../../../images/folder.svg');
-const downArrowUrl = require('../../../images/down-arrow.svg');
+import { withTranslation } from 'react-i18next';
+
 import ConnectedFileNode from './FileNode';
+
+import DownArrowIcon from '../../../images/down-filled-triangle.svg';
 
 class Sidebar extends React.Component {
   constructor(props) {
     super(props);
     this.resetSelectedFile = this.resetSelectedFile.bind(this);
     this.toggleProjectOptions = this.toggleProjectOptions.bind(this);
+    this.onBlurComponent = this.onBlurComponent.bind(this);
+    this.onFocusComponent = this.onFocusComponent.bind(this);
+
+    this.state = {
+      isFocused: false,
+    };
+  }
+
+  onBlurComponent() {
+    this.setState({ isFocused: false });
+    setTimeout(() => {
+      if (!this.state.isFocused) {
+        this.props.closeProjectOptions();
+      }
+    }, 200);
+  }
+
+  onFocusComponent() {
+    this.setState({ isFocused: true });
   }
 
   resetSelectedFile() {
@@ -22,73 +42,131 @@ class Sidebar extends React.Component {
     if (this.props.projectOptionsVisible) {
       this.props.closeProjectOptions();
     } else {
-      this.refs.sidebarOptions.focus();
+      this.sidebarOptions.focus();
       this.props.openProjectOptions();
     }
   }
 
+  userCanEditProject() {
+    let canEdit;
+    if (!this.props.owner) {
+      canEdit = true;
+    } else if (this.props.user.authenticated && this.props.owner.id === this.props.user.id) {
+      canEdit = true;
+    } else {
+      canEdit = false;
+    }
+    return canEdit;
+  }
+
   render() {
+    const canEditProject = this.userCanEditProject();
     const sidebarClass = classNames({
-      sidebar: true,
+      'sidebar': true,
       'sidebar--contracted': !this.props.isExpanded,
-      'sidebar--project-options': this.props.projectOptionsVisible
+      'sidebar--project-options': this.props.projectOptionsVisible,
+      'sidebar--cant-edit': !canEditProject
     });
+    const rootFile = this.props.files.filter(file => file.name === 'root')[0];
 
     return (
-      <nav className={sidebarClass} title="file-navigation" role="navigation">
-        <div className="sidebar__header" onContextMenu={this.toggleProjectOptions}>
+      <section className={sidebarClass}>
+        <header className="sidebar__header" onContextMenu={this.toggleProjectOptions}>
           <h3 className="sidebar__title">
-            <span className="sidebar__folder-icon">
-              <InlineSVG src={folderUrl} />
-            </span>
-            <span>project-folder</span>
+            <span>{this.props.t('Sidebar.Title')}</span>
           </h3>
           <div className="sidebar__icons">
             <button
-              aria-label="add file or folder"
+              aria-label={this.props.t('Sidebar.ToggleARIA')}
               className="sidebar__add"
               tabIndex="0"
-              ref="sidebarOptions"
+              ref={(element) => { this.sidebarOptions = element; }}
               onClick={this.toggleProjectOptions}
-              onBlur={() => setTimeout(this.props.closeProjectOptions, 200)}
+              onBlur={this.onBlurComponent}
+              onFocus={this.onFocusComponent}
             >
-              <InlineSVG src={downArrowUrl} />
+              <DownArrowIcon focusable="false" aria-hidden="true" />
             </button>
             <ul className="sidebar__project-options">
               <li>
-                <a aria-label="add folder" onClick={this.props.newFolder} >
-                  Add folder
-                </a>
+                <button
+                  aria-label={this.props.t('Sidebar.AddFolderARIA')}
+                  onClick={() => {
+                    this.props.newFolder(rootFile.id);
+                    setTimeout(this.props.closeProjectOptions, 0);
+                  }}
+                  onBlur={this.onBlurComponent}
+                  onFocus={this.onFocusComponent}
+                >
+                  {this.props.t('Sidebar.AddFolder')}
+                </button>
               </li>
               <li>
-                <a aria-label="add file" onClick={this.props.newFile} >
-                  Add file
-                </a>
+                <button
+                  aria-label={this.props.t('Sidebar.AddFileARIA')}
+                  onClick={() => {
+                    this.props.newFile(rootFile.id);
+                    setTimeout(this.props.closeProjectOptions, 0);
+                  }}
+                  onBlur={this.onBlurComponent}
+                  onFocus={this.onFocusComponent}
+                >
+                  {this.props.t('Sidebar.AddFile')}
+                </button>
               </li>
+              {
+                this.props.user.authenticated &&
+                <li>
+                  <button
+                    aria-label={this.props.t('Sidebar.UploadFileARIA')}
+                    onClick={() => {
+                      this.props.openUploadFileModal(rootFile.id);
+                      setTimeout(this.props.closeProjectOptions, 0);
+                    }}
+                    onBlur={this.onBlurComponent}
+                    onFocus={this.onFocusComponent}
+                  >
+                    {this.props.t('Sidebar.UploadFile')}
+                  </button>
+                </li>
+              }
             </ul>
           </div>
-        </div>
-        <ConnectedFileNode id={this.props.files.filter(file => file.name === 'root')[0].id} />
-      </nav>
+        </header>
+        <ConnectedFileNode
+          id={rootFile.id}
+          canEdit={canEditProject}
+        />
+      </section>
     );
   }
 }
 
 Sidebar.propTypes = {
-  files: PropTypes.array.isRequired,
+  files: PropTypes.arrayOf(PropTypes.shape({
+    name: PropTypes.string.isRequired,
+    id: PropTypes.string.isRequired
+  })).isRequired,
   setSelectedFile: PropTypes.func.isRequired,
   isExpanded: PropTypes.bool.isRequired,
   projectOptionsVisible: PropTypes.bool.isRequired,
   newFile: PropTypes.func.isRequired,
-  showFileOptions: PropTypes.func.isRequired,
-  hideFileOptions: PropTypes.func.isRequired,
-  deleteFile: PropTypes.func.isRequired,
-  showEditFileName: PropTypes.func.isRequired,
-  hideEditFileName: PropTypes.func.isRequired,
-  updateFileName: PropTypes.func.isRequired,
   openProjectOptions: PropTypes.func.isRequired,
   closeProjectOptions: PropTypes.func.isRequired,
-  newFolder: PropTypes.func.isRequired
+  newFolder: PropTypes.func.isRequired,
+  openUploadFileModal: PropTypes.func.isRequired,
+  owner: PropTypes.shape({
+    id: PropTypes.string
+  }),
+  user: PropTypes.shape({
+    id: PropTypes.string,
+    authenticated: PropTypes.bool.isRequired
+  }).isRequired,
+  t: PropTypes.func.isRequired,
 };
 
-export default Sidebar;
+Sidebar.defaultProps = {
+  owner: undefined
+};
+
+export default withTranslation()(Sidebar);
