@@ -1,60 +1,46 @@
-import React, { PropTypes } from 'react';
-import { Link, browserHistory } from 'react-router';
-import * as UserActions from '../actions';
+
+import PropTypes from 'prop-types';
+import React from 'react';
+import { Link } from 'react-router';
+import classNames from 'classnames';
 import { bindActionCreators } from 'redux';
 import { reduxForm } from 'redux-form';
+import { Helmet } from 'react-helmet';
+import { withTranslation } from 'react-i18next';
+import * as UserActions from '../actions';
 import ResetPasswordForm from '../components/ResetPasswordForm';
-import classNames from 'classnames';
-import InlineSVG from 'react-inlinesvg';
-const exitUrl = require('../../../images/exit.svg');
-const logoUrl = require('../../../images/p5js-logo.svg');
+import { validateResetPassword } from '../../../utils/reduxFormUtils';
+import Nav from '../../../components/Nav';
 
-class ResetPasswordView extends React.Component {
-  constructor(props) {
-    super(props);
-    this.gotoHomePage = this.gotoHomePage.bind(this);
-  }
-
-  componentWillMount() {
-    this.props.resetPasswordReset();
-  }
-
-  gotoHomePage() {
-    browserHistory.push('/');
-  }
-
-  render() {
-    const resetPasswordClass = classNames({
-      'reset-password': true,
-      'reset-password--submitted': this.props.user.resetPasswordInitiate,
-      'form-container': true
-    });
-    return (
+function ResetPasswordView(props) {
+  const resetPasswordClass = classNames({
+    'reset-password': true,
+    'reset-password--submitted': props.user.resetPasswordInitiate,
+    'form-container': true,
+    'user': true
+  });
+  return (
+    <div className="reset-password-container">
+      <Nav layout="dashboard" />
       <div className={resetPasswordClass}>
-        <div className="form-container__header">
-          <button className="form-container__logo-button" onClick={this.gotoHomePage}>
-            <InlineSVG src={logoUrl} alt="p5js Logo" />
-          </button>
-          <button className="form-container__exit-button" onClick={this.gotoHomePage}>
-            <InlineSVG src={exitUrl} alt="Close ResetPassword Page" />
-          </button>
-        </div>
+        <Helmet>
+          <title>{props.t('ResetPasswordView.Title')}</title>
+        </Helmet>
         <div className="form-container__content">
-          <h2 className="form-container__title">Reset Your Password</h2>
-          <ResetPasswordForm {...this.props} />
+          <h2 className="form-container__title">{props.t('ResetPasswordView.Reset')}</h2>
+          <ResetPasswordForm {...props} />
           <p className="reset-password__submitted">
-            Your password reset email should arrive shortly. If you don't see it, check
-            in your spam folder as sometimes it can end up there.
+            {props.t('ResetPasswordView.Submitted')}
           </p>
           <p className="form__navigation-options">
-            <Link className="form__login-button" to="/login">Log In</Link>
-            &nbsp;or&nbsp;
-            <Link className="form__signup-button" to="/signup">Sign Up</Link>
+            <Link className="form__login-button" to="/login">{props.t('ResetPasswordView.Login')}</Link>
+            &nbsp;{props.t('ResetPasswordView.LoginOr')}&nbsp;
+            <Link className="form__signup-button" to="/signup">{props.t('ResetPasswordView.SignUp')}</Link>
           </p>
         </div>
       </div>
-    );
-  }
+    </div>
+  );
 }
 
 ResetPasswordView.propTypes = {
@@ -62,6 +48,7 @@ ResetPasswordView.propTypes = {
   user: PropTypes.shape({
     resetPasswordInitiate: PropTypes.bool
   }).isRequired,
+  t: PropTypes.func.isRequired
 };
 
 function mapStateToProps(state) {
@@ -74,16 +61,8 @@ function mapDispatchToProps(dispatch) {
   return bindActionCreators(UserActions, dispatch);
 }
 
-function validate(formProps) {
-  const errors = {};
-  if (!formProps.email) {
-    errors.email = 'Please enter an email';
-  }
-  return errors;
-}
-
-export default reduxForm({
+export default withTranslation()(reduxForm({
   form: 'reset-password',
   fields: ['email'],
-  validate
-}, mapStateToProps, mapDispatchToProps)(ResetPasswordView);
+  validate: validateResetPassword
+}, mapStateToProps, mapDispatchToProps)(ResetPasswordView));
