@@ -1,57 +1,75 @@
-import React, { PropTypes } from 'react';
+import PropTypes from 'prop-types';
+import React from 'react';
 import { bindActionCreators } from 'redux';
-import * as UserActions from '../actions';
-import { reduxForm } from 'redux-form';
-import SignupForm from '../components/SignupForm';
-import axios from 'axios';
 import { Link, browserHistory } from 'react-router';
-import InlineSVG from 'react-inlinesvg';
-const exitUrl = require('../../../images/exit.svg');
-const logoUrl = require('../../../images/p5js-logo.svg');
+import { Helmet } from 'react-helmet';
+import { reduxForm } from 'redux-form';
+import { withTranslation } from 'react-i18next';
+import * as UserActions from '../actions';
+import SignupForm from '../components/SignupForm';
+import apiClient from '../../../utils/apiClient';
+import { validateSignup } from '../../../utils/reduxFormUtils';
+import SocialAuthButton from '../components/SocialAuthButton';
+import Nav from '../../../components/Nav';
+import ResponsiveForm from '../components/ResponsiveForm';
+
 
 class SignupView extends React.Component {
-  constructor(props) {
-    super(props);
-    this.closeSignupPage = this.closeSignupPage.bind(this);
-    this.gotoHomePage = this.gotoHomePage.bind(this);
-  }
-
-  closeSignupPage() {
-    browserHistory.push(this.props.previousPath);
-  }
-
-  gotoHomePage() {
+  gotoHomePage = () => {
     browserHistory.push('/');
   }
 
   render() {
+    if (this.props.user.authenticated) {
+      this.gotoHomePage();
+      return null;
+    }
     return (
-      <div className="form-container">
-        <div className="form-container__header">
-          <button className="form-container__logo-button" onClick={this.gotoHomePage}>
-            <InlineSVG src={logoUrl} alt="p5js Logo" />
-          </button>
-          <button className="form-container__exit-button" onClick={this.closeSignupPage}>
-            <InlineSVG src={exitUrl} alt="Close Signup Page" />
-          </button>
-        </div>
-        <div className="form-container__content">
-          <h2 className="form-container__title">Sign Up</h2>
-          <SignupForm {...this.props} />
-          <p className="form__navigation-options">
-            Already have an account?&nbsp;
-            <Link className="form__login-button" to="/login">Log In</Link>
-          </p>
-        </div>
+      <div className="signup">
+        <Nav layout="dashboard" />
+        <main className="form-container">
+          <Helmet>
+            <title>{this.props.t('SignupView.Title')}</title>
+          </Helmet>
+          <div className="form-container__content">
+            <h2 className="form-container__title">{this.props.t('SignupView.Description')}</h2>
+            <SignupForm {...this.props} />
+            <h2 className="form-container__divider">{this.props.t('SignupView.Or')}</h2>
+            <div className="form-container__stack">
+              <SocialAuthButton service={SocialAuthButton.services.github} />
+              <SocialAuthButton service={SocialAuthButton.services.google} />
+            </div>
+            <p className="form__navigation-options">
+              {this.props.t('SignupView.AlreadyHave')}
+              <Link className="form__login-button" to="/login">{this.props.t('SignupView.Login')}</Link>
+            </p>
+          </div>
+        </main>
       </div>
     );
   }
 }
 
+function asyncErrorsSelector(formName, state) {
+  const form = state.form[formName];
+  if (!form) {
+    return {};
+  }
+
+  const fieldNames = Object.keys(form).filter(key => !key.startsWith('_'));
+  return fieldNames.reduce((asyncErrors, fieldName) => {
+    if (form[fieldName].asyncError) {
+      return { ...asyncErrors, [fieldName]: form[fieldName].asyncError };
+    }
+    return asyncErrors;
+  }, {});
+}
+
 function mapStateToProps(state) {
   return {
     user: state.user,
-    previousPath: state.ide.previousPath
+    previousPath: state.ide.previousPath,
+    asyncErrors: asyncErrorsSelector('signup', state)
   };
 }
 
@@ -60,52 +78,29 @@ function mapDispatchToProps(dispatch) {
 }
 
 function asyncValidate(formProps, dispatch, props) {
-  const fieldToValidate = props.form._active;
-  if (fieldToValidate) {
-    const queryParams = {};
-    queryParams[fieldToValidate] = formProps[fieldToValidate];
-    queryParams.check_type = fieldToValidate;
-    return axios.get('/api/signup/duplicate_check', { params: queryParams })
-      .then(response => {
-        if (response.data.exists) {
-          const error = {};
-          error[fieldToValidate] = response.data.message;
-          throw error;
-        }
-      });
-  }
-  return Promise.resolve(true).then(() => {});
-}
-
-function validate(formProps) {
   const errors = {};
-
-  if (!formProps.username) {
-    errors.username = 'Please enter a username.';
-  } else if (!formProps.username.match(/^.{1,20}$/)) {
-    errors.username = 'Username must be less than 20 characters.';
-  } else if (!formProps.username.match(/^[a-zA-Z0-9._-]{1,20}$/)) {
-    errors.username = 'Username must only consist of numbers, letters, periods, dashes, and underscores.';
-  }
-
-  if (!formProps.email) {
-    errors.email = 'Please enter an email.';
-  } else if (!formProps.email.match(/^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/i)) {
-    errors.email = 'Please enter a valid email address.';
-  }
-
-  if (!formProps.password) {
-    errors.password = 'Please enter a password';
-  }
-  if (!formProps.confirmPassword) {
-    errors.confirmPassword = 'Please enter a password confirmation';
-  }
-
-  if (formProps.password !== formProps.confirmPassword) {
-    errors.password = 'Passwords must match';
-  }
-
-  return errors;
+  return Promise.resolve(true)
+    .then(() => {
+      const fieldToValidate = props.form._active;
+      if (fieldToValidate) {
+        const queryParams = {};
+        queryParams[fieldToValidate] = formProps[fieldToValidate];
+        queryParams.check_type = fieldToValidate;
+        return apiClient.get('/signup/duplicate_check', { params: queryParams })
+          .then((response) => {
+            if (response.data.exists) {
+              errors[fieldToValidate] = response.data.message;
+            }
+          });
+      }
+      return null;
+    })
+    .then(() => {
+      const err = { ...errors, ...props.asyncErrors };
+      if (Object.keys(err).length > 0) {
+        throw err;
+      }
+    });
 }
 
 function onSubmitFail(errors) {
@@ -113,14 +108,24 @@ function onSubmitFail(errors) {
 }
 
 SignupView.propTypes = {
-  previousPath: PropTypes.string.isRequired
+  previousPath: PropTypes.string.isRequired,
+  user: PropTypes.shape({
+    authenticated: PropTypes.bool
+  }),
+  t: PropTypes.func.isRequired,
 };
 
-export default reduxForm({
+SignupView.defaultProps = {
+  user: {
+    authenticated: false
+  },
+};
+
+export default withTranslation()(reduxForm({
   form: 'signup',
   fields: ['username', 'email', 'password', 'confirmPassword'],
   onSubmitFail,
-  validate,
+  validate: validateSignup,
   asyncValidate,
   asyncBlurFields: ['username', 'email']
-}, mapStateToProps, mapDispatchToProps)(SignupView);
+}, mapStateToProps, mapDispatchToProps)(SignupView));
