@@ -2,18 +2,59 @@ import Express from 'express';
 import mongoose from 'mongoose';
 import bodyParser from 'body-parser';
 import cookieParser from 'cookie-parser';
+import cors from 'cors';
 import session from 'express-session';
-const MongoStore = require('connect-mongo')(session);
+import connectMongo from 'connect-mongo';
 import passport from 'passport';
 import path from 'path';
+import basicAuth from 'express-basic-auth';
 
 // Webpack Requirements
 import webpack from 'webpack';
-import config from '../webpack.config.dev';
 import webpackDevMiddleware from 'webpack-dev-middleware';
 import webpackHotMiddleware from 'webpack-hot-middleware';
+import config from '../webpack/config.dev';
+
+// Import all required modules
+import api from './routes/api.routes';
+import users from './routes/user.routes';
+import sessions from './routes/session.routes';
+import projects from './routes/project.routes';
+import files from './routes/file.routes';
+import collections from './routes/collection.routes';
+import aws from './routes/aws.routes';
+import serverRoutes from './routes/server.routes';
+import embedRoutes from './routes/embed.routes';
+import assetRoutes from './routes/asset.routes';
+import passportRoutes from './routes/passport.routes';
+import { requestsOfTypeJSON } from './utils/requestsOfType';
+
+import { renderIndex } from './views/index';
+import { get404Sketch } from './views/404Page';
 
 const app = new Express();
+const MongoStore = connectMongo(session);
+
+app.get('/health', (req, res) => res.json({ success: true }));
+
+// For basic auth, in setting up beta editor
+if (process.env.BASIC_USERNAME && process.env.BASIC_PASSWORD) {
+  app.use(basicAuth({
+    users: {
+      [process.env.BASIC_USERNAME]: process.env.BASIC_PASSWORD
+    },
+    challenge: true
+  }));
+}
+
+const allowedCorsOrigins = [
+  /p5js\.org$/,
+];
+
+// to allow client-only development
+if (process.env.CORS_ALLOW_LOCALHOST === 'true') {
+  allowedCorsOrigins.push(/localhost/);
+}
 
 // Run Webpack dev server in development mode
 if (process.env.NODE_ENV === 'development') {
@@ -22,24 +63,37 @@ if (process.env.NODE_ENV === 'development') {
   app.use(webpackHotMiddleware(compiler));
 }
 
-// Import all required modules
-import serverConfig from './config';
-import users from './routes/user.routes';
-import sessions from './routes/session.routes';
-import projects from './routes/project.routes';
-import files from './routes/file.routes';
-import aws from './routes/aws.routes';
-import serverRoutes from './routes/server.routes';
-import embedRoutes from './routes/embed.routes';
+const mongoConnectionString = process.env.MONGO_URL;
+app.set('trust proxy', true);
 
-import { renderIndex } from './views/index';
-import { get404Sketch } from './views/404Page';
+// Enable Cross-Origin Resource Sharing (CORS) for all origins
+const corsMiddleware = cors({
+  credentials: true,
+  origin: allowedCorsOrigins,
+});
+app.use(corsMiddleware);
+// Enable pre-flight OPTIONS route for all end-points
+app.options('*', corsMiddleware);
 
 // Body parser, cookie parser, sessions, serve public assets
+app.use(
+  '/locales',
+  Express.static(
+    path.resolve(__dirname, '../dist/static/locales'),
+    {
+      // Browsers must revalidate for changes to the locale files
+      // It doesn't actually mean "don't cache this file"
+      // See: https://jakearchibald.com/2016/caching-best-practices/
+      setHeaders: res => res.setHeader('Cache-Control', 'no-cache')
+    }
+  )
+);
+app.use(Express.static(path.resolve(__dirname, '../dist/static'), {
+  maxAge: process.env.STATIC_MAX_AGE || (process.env.NODE_ENV === 'production' ? '1d' : '0')
+}));
 
-app.use(Express.static(path.resolve(__dirname, '../static')));
-app.use(bodyParser.urlencoded({ extended: true }));
-app.use(bodyParser.json());
+app.use(bodyParser.urlencoded({ limit: '50mb', extended: true }));
+app.use(bodyParser.json({ limit: '50mb' }));
 app.use(cookieParser());
 app.use(session({
   resave: true,
@@ -52,34 +106,45 @@ app.use(session({
     secure: false,
   },
   store: new MongoStore({
-    url: process.env.MONGO_URL,
+    url: mongoConnectionString,
     autoReconnect: true
   })
 }));
+
 app.use(passport.initialize());
 app.use(passport.session());
-app.use('/api', users);
-app.use('/api', sessions);
-app.use('/api', projects);
-app.use('/api', files);
-app.use('/api', aws);
+app.use('/api/v1', requestsOfTypeJSON(), api);
+app.use('/editor', requestsOfTypeJSON(), users);
+app.use('/editor', requestsOfTypeJSON(), sessions);
+app.use('/editor', requestsOfTypeJSON(), files);
+app.use('/editor', requestsOfTypeJSON(), projects);
+app.use('/editor', requestsOfTypeJSON(), aws);
+app.use('/editor', requestsOfTypeJSON(), collections);
+
+// This is a temporary way to test access via Personal Access Tokens
+// Sending a valid username:<personal-access-token> combination will
+// return the user's information.
+app.get(
+  '/api/v1/auth/access-check',
+  passport.authenticate('basic', { session: false }), (req, res) => res.json(req.user)
+);
+
 // this is supposed to be TEMPORARY -- until i figure out
 // isomorphic rendering
 app.use('/', serverRoutes);
 
+app.use(assetRoutes);
+
 app.use('/', embedRoutes);
-app.get('/auth/github', passport.authenticate('github'));
-app.get('/auth/github/callback', passport.authenticate('github', { failureRedirect: '/login' }), (req, res) => {
-  res.redirect('/');
-});
+app.use('/', passportRoutes);
 
 // configure passport
 require('./config/passport');
-// const passportConfig = require('./config/passport');
 
 // Connect to MongoDB
-// mongoose.connect(process.env.MONGODB_URI || process.env.MONGOLAB_URI);
-mongoose.connect(serverConfig.mongoURL);
+mongoose.Promise = global.Promise;
+mongoose.connect(mongoConnectionString, { useNewUrlParser: true, useUnifiedTopology: true });
+mongoose.set('useCreateIndex', true);
 mongoose.connection.on('error', () => {
   console.error('MongoDB Connection Error. Please make sure that MongoDB is running.');
   process.exit(1);
@@ -88,6 +153,17 @@ mongoose.connection.on('error', () => {
 app.get('/', (req, res) => {
   res.sendFile(renderIndex());
 });
+
+// Handle API errors
+app.use('/api', (error, req, res, next) => {
+  if (error && error.code && !res.headersSent) {
+    res.status(error.code).json({ error: error.message });
+    return;
+  }
+
+  next(error);
+});
+
 
 // Handle missing routes.
 app.get('*', (req, res) => {
@@ -104,9 +180,9 @@ app.get('*', (req, res) => {
 });
 
 // start app
-app.listen(serverConfig.port, (error) => {
+app.listen(process.env.PORT, (error) => {
   if (!error) {
-    console.log(`p5js web editor is running on port: ${serverConfig.port}!`); // eslint-disable-line
+    console.log(`p5.js Web Editor is running on port: ${process.env.PORT}!`); // eslint-disable-line
   }
 });
 
