@@ -3,7 +3,7 @@ import { createDevTools } from 'redux-devtools';
 import LogMonitor from 'redux-devtools-log-monitor';
 import DockMonitor from 'redux-devtools-dock-monitor';
 
-export default createDevTools(
+const devTools = (
   <DockMonitor
     toggleVisibilityKey="ctrl-h"
     changePositionKey="ctrl-w"
@@ -11,3 +11,5 @@ export default createDevTools(
     <LogMonitor />
   </DockMonitor>
 );
+
+export default createDevTools(devTools);
