@@ -1,55 +1,32 @@
+import apiClient from '../../../utils/apiClient';
 import * as ActionTypes from '../../../constants';
-import axios from 'axios';
-import { showErrorModal, setPreviousPath } from './ide';
-import { resetProject } from './project';
+import { startLoader, stopLoader } from './loader';
 
-const ROOT_URL = location.href.indexOf('localhost') > 0 ? 'http://localhost:8000/api' : '/api';
-
+// eslint-disable-next-line
 export function getProjects(username) {
   return (dispatch) => {
+    dispatch(startLoader());
     let url;
     if (username) {
-      url = `${ROOT_URL}/${username}/projects`;
+      url = `/${username}/projects`;
     } else {
-      url = `${ROOT_URL}/projects`;
+      url = '/projects';
     }
-    axios.get(url, { withCredentials: true })
-      .then(response => {
+    apiClient.get(url)
+      .then((response) => {
         dispatch({
           type: ActionTypes.SET_PROJECTS,
           projects: response.data
         });
+        dispatch(stopLoader());
       })
-      .catch(response => dispatch({
-        type: ActionTypes.ERROR,
-        error: response.data
-      }));
-  };
-}
-
-export function deleteProject(id) {
-  return (dispatch, getState) => {
-    axios.delete(`${ROOT_URL}/projects/${id}`, { withCredentials: true })
-      .then(() => {
-        const state = getState();
-        if (id === state.project.id) {
-          dispatch(resetProject());
-          dispatch(setPreviousPath('/'));
-        }
+      .catch((error) => {
+        const { response } = error;
         dispatch({
-          type: ActionTypes.DELETE_PROJECT,
-          id
+          type: ActionTypes.ERROR,
+          error: response.data
         });
-      })
-      .catch(response => {
-        if (response.status === 403) {
-          dispatch(showErrorModal('staleSession'));
-        } else {
-          dispatch({
-            type: ActionTypes.ERROR,
-            error: response.data
-          });
-        }
+        dispatch(stopLoader());
       });
   };
 }
