@@ -1,5 +1,9 @@
-import React, { PropTypes } from 'react';
+import PropTypes from 'prop-types';
+import React from 'react';
+import { withTranslation } from 'react-i18next';
 import { domOnlyProps } from '../../../utils/reduxFormUtils';
+
+import Button from '../../../common/Button';
 
 class NewFileForm extends React.Component {
   constructor(props) {
@@ -7,8 +11,15 @@ class NewFileForm extends React.Component {
     this.createFile = this.props.createFile.bind(this);
   }
 
+  componentDidMount() {
+    this.fileName.focus();
+  }
+
   render() {
-    const { fields: { name }, handleSubmit } = this.props;
+    const {
+      fields: { name },
+      handleSubmit,
+    } = this.props;
     return (
       <form
         className="new-file-form"
@@ -17,16 +28,29 @@ class NewFileForm extends React.Component {
           handleSubmit(this.createFile)(data);
         }}
       >
-        <label className="new-file-form__name-label" htmlFor="name">Name:</label>
-        <input
-          className="new-file-form__name-input"
-          id="name"
-          type="text"
-          placeholder="Name"
-          {...domOnlyProps(name)}
-        />
-        <input type="submit" value="Add File" aria-label="add file" />
-        {name.touched && name.error && <span className="form-error">{name.error}</span>}
+        <div className="new-file-form__input-wrapper">
+          <label className="new-file-form__name-label" htmlFor="name">
+            Name:
+          </label>
+          <input
+            className="new-file-form__name-input"
+            id="name"
+            type="text"
+            placeholder={this.props.t('NewFileForm.Placeholder')}
+            maxLength="128"
+            {...domOnlyProps(name)}
+            ref={(element) => {
+              this.fileName = element;
+            }}
+          />
+          <Button
+            type="submit"
+          >{this.props.t('NewFileForm.AddFileSubmit')}
+          </Button>
+        </div>
+        {name.touched && name.error && (
+          <span className="form-error">{name.error}</span>
+        )}
       </form>
     );
   }
@@ -38,7 +62,8 @@ NewFileForm.propTypes = {
   }).isRequired,
   handleSubmit: PropTypes.func.isRequired,
   createFile: PropTypes.func.isRequired,
-  focusOnModal: PropTypes.func.isRequired
+  focusOnModal: PropTypes.func.isRequired,
+  t: PropTypes.func.isRequired
 };
 
-export default NewFileForm;
+export default withTranslation()(NewFileForm);
