@@ -1,23 +1,31 @@
-import React, { PropTypes } from 'react';
+import PropTypes from 'prop-types';
+import React from 'react';
+import Helmet from 'react-helmet';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import PreviewFrame from '../components/PreviewFrame';
+import PreviewNav from '../../../components/PreviewNav';
 import { getHTMLFile, getJSFiles, getCSSFiles } from '../reducers/files';
 import * as ProjectActions from '../actions/project';
 
-
 class FullView extends React.Component {
   componentDidMount() {
-    this.props.getProject(this.props.params.project_id);
+    this.props.getProject(this.props.params.project_id, this.props.params.username);
   }
+
+  ident = () => {}
 
   render() {
     return (
       <div className="fullscreen-preview">
-        <h1 className="fullscreen-preview__title">
-          {this.props.project.name} {this.props.project.owner ? `by ${this.props.project.owner.username}` : ''}
-        </h1>
-        <div className="fullscreen-preview__frame-wrapper">
+        <Helmet>
+          <title>{this.props.project.name}</title>
+        </Helmet>
+        <PreviewNav
+          owner={{ username: this.props.project.owner ? `${this.props.project.owner.username}` : '' }}
+          project={{ name: this.props.project.name, id: this.props.params.project_id }}
+        />
+        <main className="preview-frame-holder">
           <PreviewFrame
             htmlFile={this.props.htmlFile}
             jsFiles={this.props.jsFiles}
@@ -28,8 +36,19 @@ class FullView extends React.Component {
             }
             fullView
             isPlaying
+            isAccessibleOutputPlaying={false}
+            textOutput={false}
+            gridOutput={false}
+            soundOutput={false}
+            dispatchConsoleEvent={this.ident}
+            endSketchRefresh={this.ident}
+            previewIsRefreshing={false}
+            setBlobUrl={this.ident}
+            stopSketch={this.ident}
+            expandConsole={this.ident}
+            clearConsole={this.ident}
           />
-        </div>
+        </main>
       </div>
     );
   }
@@ -37,19 +56,36 @@ class FullView extends React.Component {
 
 FullView.propTypes = {
   params: PropTypes.shape({
-    project_id: PropTypes.string
-  }),
+    project_id: PropTypes.string,
+    username: PropTypes.string
+  }).isRequired,
   project: PropTypes.shape({
     name: PropTypes.string,
     owner: PropTypes.shape({
       username: PropTypes.string
     })
   }).isRequired,
-  htmlFile: PropTypes.object,
-  jsFiles: PropTypes.array,
-  cssFiles: PropTypes.array,
+  htmlFile: PropTypes.shape({
+    id: PropTypes.string.isRequired,
+    content: PropTypes.string.isRequired,
+    name: PropTypes.string.isRequired
+  }).isRequired,
+  jsFiles: PropTypes.arrayOf(PropTypes.shape({
+    id: PropTypes.string.isRequired,
+    content: PropTypes.string.isRequired,
+    name: PropTypes.string.isRequired
+  })).isRequired,
+  cssFiles: PropTypes.arrayOf(PropTypes.shape({
+    id: PropTypes.string.isRequired,
+    content: PropTypes.string.isRequired,
+    name: PropTypes.string.isRequired
+  })).isRequired,
   getProject: PropTypes.func.isRequired,
-  files: PropTypes.array
+  files: PropTypes.arrayOf(PropTypes.shape({
+    id: PropTypes.string.isRequired,
+    content: PropTypes.string.isRequired,
+    name: PropTypes.string.isRequired
+  })).isRequired,
 };
 
 function mapStateToProps(state) {
