@@ -1,21 +1,25 @@
-import React, { PropTypes } from 'react';
+import PropTypes from 'prop-types';
+import React from 'react';
+import { bindActionCreators } from 'redux';
+import { connect } from 'react-redux';
+import { withRouter } from 'react-router';
+import { withTranslation } from 'react-i18next';
+import { Helmet } from 'react-helmet';
+import SplitPane from 'react-split-pane';
 import Editor from '../components/Editor';
 import Sidebar from '../components/Sidebar';
 import PreviewFrame from '../components/PreviewFrame';
 import Toolbar from '../components/Toolbar';
-import TextOutput from '../components/TextOutput';
-import Preferences from '../components/Preferences';
+import Preferences from '../components/Preferences/index';
 import NewFileModal from '../components/NewFileModal';
 import NewFolderModal from '../components/NewFolderModal';
+import UploadFileModal from '../components/UploadFileModal';
 import ShareModal from '../components/ShareModal';
 import KeyboardShortcutModal from '../components/KeyboardShortcutModal';
 import ErrorModal from '../components/ErrorModal';
 import Nav from '../../../components/Nav';
 import Console from '../components/Console';
 import Toast from '../components/Toast';
-import { bindActionCreators } from 'redux';
-import { connect } from 'react-redux';
-import { withRouter } from 'react-router';
 import * as FileActions from '../actions/files';
 import * as IDEActions from '../actions/ide';
 import * as ProjectActions from '../actions/project';
@@ -25,41 +29,82 @@ import * as UserActions from '../../User/actions';
 import * as ToastActions from '../actions/toast';
 import * as ConsoleActions from '../actions/console';
 import { getHTMLFile } from '../reducers/files';
-import SplitPane from 'react-split-pane';
 import Overlay from '../../App/components/Overlay';
-import SketchList from '../components/SketchList';
 import About from '../components/About';
+import AddToCollectionList from '../components/AddToCollectionList';
+import Feedback from '../components/Feedback';
+import { CollectionSearchbar } from '../components/Searchbar';
+import { getIsUserOwner } from '../selectors/users';
+
+
+function getTitle(props) {
+  const { id } = props.project;
+  return id ? `p5.js Web Editor | ${props.project.name}` : 'p5.js Web Editor';
+}
+
+function warnIfUnsavedChanges(props) {
+  // eslint-disable-line
+  const { route } = props.route;
+  if (
+    route &&
+    route.action === 'PUSH' &&
+      (route.pathname === '/login' || route.pathname === '/signup')
+  ) {
+    // don't warn
+    props.persistState();
+    window.onbeforeunload = null;
+  } else if (
+    route &&
+    (props.location.pathname === '/login' ||
+      props.location.pathname === '/signup')
+  ) {
+    // don't warn
+    props.persistState();
+    window.onbeforeunload = null;
+  } else if (props.ide.unsavedChanges) {
+    if (!window.confirm(props.t('Nav.WarningUnsavedChanges'))) {
+      return false;
+    }
+    props.setUnsavedChanges(false);
+    return true;
+  }
+  return true;
+}
 
 class IDEView extends React.Component {
   constructor(props) {
     super(props);
-    this._handleConsolePaneOnDragFinished = this._handleConsolePaneOnDragFinished.bind(this);
-    this._handleSidebarPaneOnDragFinished = this._handleSidebarPaneOnDragFinished.bind(this);
     this.handleGlobalKeydown = this.handleGlobalKeydown.bind(this);
-    this.warnIfUnsavedChanges = this.warnIfUnsavedChanges.bind(this);
+
+    this.state = {
+      consoleSize: props.ide.consoleIsExpanded ? 150 : 29,
+      sidebarSize: props.ide.sidebarIsExpanded ? 160 : 20,
+    };
   }
 
   componentDidMount() {
+    // If page doesn't reload after Sign In then we need
+    // to force cleared state to be cleared
+    this.props.clearPersistedState();
+
     this.props.stopSketch();
     if (this.props.params.project_id) {
-      const id = this.props.params.project_id;
+      const { project_id: id, username } = this.props.params;
       if (id !== this.props.project.id) {
-        this.props.getProject(id);
+        this.props.getProject(id, username);
       }
     }
-
-    this.consoleSize = this.props.ide.consoleIsExpanded ? 150 : 29;
-    this.sidebarSize = this.props.ide.sidebarIsExpanded ? 160 : 20;
-    this.forceUpdate();
 
     this.isMac = navigator.userAgent.toLowerCase().indexOf('mac') !== -1;
     document.addEventListener('keydown', this.handleGlobalKeydown, false);
 
-    this.props.router.setRouteLeaveHook(this.props.route, (route) => this.warnIfUnsavedChanges(route));
+    this.props.router.setRouteLeaveHook(
+      this.props.route,
+      this.handleUnsavedChanges
+    );
 
-    window.onbeforeunload = () => this.warnIfUnsavedChanges();
+    window.onbeforeunload = this.handleUnsavedChanges;
 
-    document.body.className = this.props.preferences.theme;
     this.autosaveInterval = null;
   }
 
@@ -67,33 +112,44 @@ class IDEView extends React.Component {
     if (nextProps.location !== this.props.location) {
       this.props.setPreviousPath(this.props.location.pathname);
     }
-  }
 
-  componentWillUpdate(nextProps) {
     if (this.props.ide.consoleIsExpanded !== nextProps.ide.consoleIsExpanded) {
-      this.consoleSize = nextProps.ide.consoleIsExpanded ? 150 : 29;
+      this.setState({
+        consoleSize: nextProps.ide.consoleIsExpanded ? 150 : 29,
+      });
     }
 
     if (this.props.ide.sidebarIsExpanded !== nextProps.ide.sidebarIsExpanded) {
-      this.sidebarSize = nextProps.ide.sidebarIsExpanded ? 160 : 20;
+      this.setState({
+        sidebarSize: nextProps.ide.sidebarIsExpanded ? 160 : 20,
+      });
     }
+  }
 
+  componentWillUpdate(nextProps) {
     if (nextProps.params.project_id && !this.props.params.project_id) {
       if (nextProps.params.project_id !== nextProps.project.id) {
         this.props.getProject(nextProps.params.project_id);
       }
     }
-
-    if (nextProps.preferences.theme !== this.props.preferences.theme) {
-      document.body.className = nextProps.preferences.theme;
-    }
   }
 
   componentDidUpdate(prevProps) {
-    if (this.isUserOwner() && this.props.project.id) {
-      if (this.props.preferences.autosave && this.props.ide.unsavedChanges && this.autosaveInterval === null && !this.props.ide.justOpenedProject) {
-        console.log('saving project in 30 seconds');
-        this.autosaveInterval = setTimeout(this.props.autosaveProject, 30000);
+    if (this.props.isUserOwner && this.props.project.id) {
+      if (
+        this.props.preferences.autosave &&
+        this.props.ide.unsavedChanges &&
+        !this.props.ide.justOpenedProject
+      ) {
+        if (
+          this.props.selectedFile.name === prevProps.selectedFile.name &&
+          this.props.selectedFile.content !== prevProps.selectedFile.content
+        ) {
+          if (this.autosaveInterval) {
+            clearTimeout(this.autosaveInterval);
+          }
+          this.autosaveInterval = setTimeout(this.props.autosaveProject, 20000);
+        }
       } else if (this.autosaveInterval && !this.props.preferences.autosave) {
         clearTimeout(this.autosaveInterval);
         this.autosaveInterval = null;
@@ -104,341 +160,293 @@ class IDEView extends React.Component {
     }
 
     if (this.props.route.path !== prevProps.route.path) {
-      this.props.router.setRouteLeaveHook(this.props.route, (route) => this.warnIfUnsavedChanges(route));
+      this.props.router.setRouteLeaveHook(this.props.route, () =>
+        warnIfUnsavedChanges(this.props));
     }
   }
-
   componentWillUnmount() {
+    document.removeEventListener('keydown', this.handleGlobalKeydown, false);
     clearTimeout(this.autosaveInterval);
     this.autosaveInterval = null;
-    this.consoleSize = undefined;
-    this.sidebarSize = undefined;
   }
-
-  isUserOwner() {
-    return this.props.project.owner && this.props.project.owner.id === this.props.user.id;
-  }
-
-  _handleConsolePaneOnDragFinished() {
-    this.consoleSize = this.refs.consolePane.state.draggedSize;
-    this.refs.consolePane.setState({
-      resized: false,
-      draggedSize: undefined,
-    });
-  }
-
-  _handleSidebarPaneOnDragFinished() {
-    this.sidebarSize = this.refs.sidebarPane.state.draggedSize;
-    this.refs.sidebarPane.setState({
-      resized: false,
-      draggedSize: undefined
-    });
-  }
-
   handleGlobalKeydown(e) {
     // 83 === s
-    if (e.keyCode === 83 && ((e.metaKey && this.isMac) || (e.ctrlKey && !this.isMac))) {
+    if (
+      e.keyCode === 83 &&
+      ((e.metaKey && this.isMac) || (e.ctrlKey && !this.isMac))
+    ) {
       e.preventDefault();
       e.stopPropagation();
-      if (this.isUserOwner() || this.props.user.authenticated && !this.props.project.owner) {
-        this.props.saveProject();
+      if (
+        this.props.isUserOwner ||
+        (this.props.user.authenticated && !this.props.project.owner)
+      ) {
+        this.props.saveProject(this.cmController.getContent());
+      } else if (this.props.user.authenticated) {
+        this.props.cloneProject();
+      } else {
+        this.props.showErrorModal('forceAuthentication');
       }
       // 13 === enter
-    } else if (e.keyCode === 13 && e.shiftKey && ((e.metaKey && this.isMac) || (e.ctrlKey && !this.isMac))) {
+    } else if (
+      e.keyCode === 13 &&
+      e.shiftKey &&
+      ((e.metaKey && this.isMac) || (e.ctrlKey && !this.isMac))
+    ) {
       e.preventDefault();
       e.stopPropagation();
       this.props.stopSketch();
-    } else if (e.keyCode === 13 && ((e.metaKey && this.isMac) || (e.ctrlKey && !this.isMac))) {
+    } else if (
+      e.keyCode === 13 &&
+      ((e.metaKey && this.isMac) || (e.ctrlKey && !this.isMac))
+    ) {
       e.preventDefault();
       e.stopPropagation();
-      this.props.clearConsole();
-      this.props.startSketchAndRefresh();
-    } else if (e.keyCode === 50 && ((e.metaKey && this.isMac) || (e.ctrlKey && !this.isMac)) && e.shiftKey) {
+      this.props.startSketch();
+      // 50 === 2
+    } else if (
+      e.keyCode === 50 &&
+      ((e.metaKey && this.isMac) || (e.ctrlKey && !this.isMac)) &&
+      e.shiftKey
+    ) {
       e.preventDefault();
-      this.props.setTextOutput(0);
-    } else if (e.keyCode === 49 && ((e.metaKey && this.isMac) || (e.ctrlKey && !this.isMac)) && e.shiftKey) {
+      this.props.setAllAccessibleOutput(false);
+      // 49 === 1
+    } else if (
+      e.keyCode === 49 &&
+      ((e.metaKey && this.isMac) || (e.ctrlKey && !this.isMac)) &&
+      e.shiftKey
+    ) {
       e.preventDefault();
-      if (this.props.preferences.textOutput === 3) {
-        this.props.preferences.textOutput = 1;
+      this.props.setAllAccessibleOutput(true);
+    } else if (
+      e.keyCode === 66 &&
+      ((e.metaKey && this.isMac) || (e.ctrlKey && !this.isMac))
+    ) {
+      e.preventDefault();
+      if (!this.props.ide.sidebarIsExpanded) {
+        this.props.expandSidebar();
       } else {
-        this.props.preferences.textOutput += 1;
+        this.props.collapseSidebar();
       }
-      this.props.setTextOutput(this.props.preferences.textOutput);
+    } else if (e.keyCode === 192 && e.ctrlKey) {
+      e.preventDefault();
+      if (this.props.ide.consoleIsExpanded) {
+        this.props.collapseConsole();
+      } else {
+        this.props.expandConsole();
+      }
+    } else if (e.keyCode === 27) {
+      if (this.props.ide.newFolderModalVisible) {
+        this.props.closeNewFolderModal();
+      } else if (this.props.ide.uploadFileModalVisible) {
+        this.props.closeUploadFileModal();
+      } else if (this.props.ide.modalIsVisible) {
+        this.props.closeNewFileModal();
+      }
     }
   }
 
-  warnIfUnsavedChanges(route) { // eslint-disable-line
-    if (route && (route.action === 'PUSH' && (route.pathname === '/login' || route.pathname === '/signup'))) {
-      // don't warn
-    } else if (route && this.props.location.pathname === '/login' || this.props.location.pathname === '/signup') {
-      // don't warn
-    } else if (this.props.ide.unsavedChanges) {
-      if (!window.confirm('Are you sure you want to leave this page? You have unsaved changes.')) {
-        return false;
-      }
-      this.props.setUnsavedChanges(false);
-      return true;
-    }
-  }
+  handleUnsavedChanges = () => warnIfUnsavedChanges(this.props);
 
   render() {
     return (
       <div className="ide">
+        <Helmet>
+          <title>{getTitle(this.props)}</title>
+        </Helmet>
         {this.props.toast.isVisible && <Toast />}
         <Nav
-          user={this.props.user}
-          newProject={this.props.newProject}
-          saveProject={this.props.saveProject}
-          exportProjectAsZip={this.props.exportProjectAsZip}
-          cloneProject={this.props.cloneProject}
-          project={this.props.project}
-          logoutUser={this.props.logoutUser}
-          stopSketch={this.props.stopSketch}
-          showShareModal={this.props.showShareModal}
-          showErrorModal={this.props.showErrorModal}
-          unsavedChanges={this.props.ide.unsavedChanges}
-          warnIfUnsavedChanges={this.warnIfUnsavedChanges}
+          warnIfUnsavedChanges={this.handleUnsavedChanges}
+          cmController={this.cmController}
         />
-        <Toolbar
-          className="Toolbar"
-          isPlaying={this.props.ide.isPlaying}
-          startSketch={this.props.startSketch}
-          stopSketch={this.props.stopSketch}
-          startTextOutput={this.props.startTextOutput}
-          stopTextOutput={this.props.stopTextOutput}
-          projectName={this.props.project.name}
-          setProjectName={this.props.setProjectName}
-          showEditProjectName={this.props.showEditProjectName}
-          hideEditProjectName={this.props.hideEditProjectName}
-          openPreferences={this.props.openPreferences}
-          preferencesIsVisible={this.props.ide.preferencesIsVisible}
-          setTextOutput={this.props.setTextOutput}
-          owner={this.props.project.owner}
-          project={this.props.project}
-          infiniteLoop={this.props.ide.infiniteLoop}
-          autorefresh={this.props.preferences.autorefresh}
-          setAutorefresh={this.props.setAutorefresh}
-          startSketchAndRefresh={this.props.startSketchAndRefresh}
-          saveProject={this.props.saveProject}
-          currentUser={this.props.user.username}
-          clearConsole={this.props.clearConsole}
-        />
-        <Preferences
-          isVisible={this.props.ide.preferencesIsVisible}
-          closePreferences={this.props.closePreferences}
-          fontSize={this.props.preferences.fontSize}
-          indentationAmount={this.props.preferences.indentationAmount}
-          setIndentation={this.props.setIndentation}
-          indentWithSpace={this.props.indentWithSpace}
-          indentWithTab={this.props.indentWithTab}
-          isTabIndent={this.props.preferences.isTabIndent}
-          setFontSize={this.props.setFontSize}
-          autosave={this.props.preferences.autosave}
-          setAutosave={this.props.setAutosave}
-          lintWarning={this.props.preferences.lintWarning}
-          setLintWarning={this.props.setLintWarning}
-          textOutput={this.props.preferences.textOutput}
-          setTextOutput={this.props.setTextOutput}
-          theme={this.props.preferences.theme}
-          setTheme={this.props.setTheme}
-        />
-        <div className="editor-preview-container">
+        <Toolbar key={this.props.project.id} />
+        {this.props.ide.preferencesIsVisible && (
+          <Overlay
+            title={this.props.t('Preferences.Settings')}
+            ariaLabel={this.props.t('Preferences.Settings')}
+            closeOverlay={this.props.closePreferences}
+          >
+            <Preferences
+              fontSize={this.props.preferences.fontSize}
+              setFontSize={this.props.setFontSize}
+              autosave={this.props.preferences.autosave}
+              linewrap={this.props.preferences.linewrap}
+              lineNumbers={this.props.preferences.lineNumbers}
+              setLineNumbers={this.props.setLineNumbers}
+              setAutosave={this.props.setAutosave}
+              setLinewrap={this.props.setLinewrap}
+              lintWarning={this.props.preferences.lintWarning}
+              setLintWarning={this.props.setLintWarning}
+              textOutput={this.props.preferences.textOutput}
+              gridOutput={this.props.preferences.gridOutput}
+              soundOutput={this.props.preferences.soundOutput}
+              setTextOutput={this.props.setTextOutput}
+              setGridOutput={this.props.setGridOutput}
+              setSoundOutput={this.props.setSoundOutput}
+              theme={this.props.preferences.theme}
+              setTheme={this.props.setTheme}
+              autocloseBracketsQuotes={this.props.preferences.autocloseBracketsQuotes}
+              setAutocloseBracketsQuotes={this.props.setAutocloseBracketsQuotes}
+            />
+          </Overlay>
+        )}
+        <main className="editor-preview-container">
           <SplitPane
             split="vertical"
-            defaultSize={this.sidebarSize}
-            ref="sidebarPane"
+            size={this.state.sidebarSize}
+            onChange={size => this.setState({ sidebarSize: size })}
             onDragFinished={this._handleSidebarPaneOnDragFinished}
             allowResize={this.props.ide.sidebarIsExpanded}
-            minSize={20}
+            minSize={125}
           >
             <Sidebar
               files={this.props.files}
               setSelectedFile={this.props.setSelectedFile}
               newFile={this.props.newFile}
               isExpanded={this.props.ide.sidebarIsExpanded}
-              showFileOptions={this.props.showFileOptions}
-              hideFileOptions={this.props.hideFileOptions}
               deleteFile={this.props.deleteFile}
-              showEditFileName={this.props.showEditFileName}
-              hideEditFileName={this.props.hideEditFileName}
               updateFileName={this.props.updateFileName}
               projectOptionsVisible={this.props.ide.projectOptionsVisible}
               openProjectOptions={this.props.openProjectOptions}
               closeProjectOptions={this.props.closeProjectOptions}
               newFolder={this.props.newFolder}
+              user={this.props.user}
+              owner={this.props.project.owner}
+              openUploadFileModal={this.props.openUploadFileModal}
+              closeUploadFileModal={this.props.closeUploadFileModal}
             />
             <SplitPane
               split="vertical"
-              defaultSize={'50%'}
-              onChange={() => (this.refs.overlay.style.display = 'block')}
-              onDragFinished={() => (this.refs.overlay.style.display = 'none')}
+              defaultSize="50%"
+              onChange={() => {
+                this.overlay.style.display = 'block';
+              }}
+              onDragFinished={() => {
+                this.overlay.style.display = 'none';
+              }}
+              resizerStyle={{
+                borderLeftWidth: '2px',
+                borderRightWidth: '2px',
+                width: '2px',
+                margin: '0px 0px',
+              }}
             >
               <SplitPane
                 split="horizontal"
                 primary="second"
-                defaultSize={this.consoleSize}
+                size={this.state.consoleSize}
                 minSize={29}
-                ref="consolePane"
-                onDragFinished={this._handleConsolePaneOnDragFinished}
+                onChange={size => this.setState({ consoleSize: size })}
                 allowResize={this.props.ide.consoleIsExpanded}
                 className="editor-preview-subpanel"
               >
-                <Editor
-                  lintWarning={this.props.preferences.lintWarning}
-                  lintMessages={this.props.editorAccessibility.lintMessages}
-                  updateLineNumber={this.props.updateLineNumber}
-                  updateLintMessage={this.props.updateLintMessage}
-                  clearLintMessage={this.props.clearLintMessage}
-                  file={this.props.selectedFile}
-                  updateFileContent={this.props.updateFileContent}
-                  fontSize={this.props.preferences.fontSize}
-                  indentationAmount={this.props.preferences.indentationAmount}
-                  isTabIndent={this.props.preferences.isTabIndent}
-                  files={this.props.files}
-                  lintMessages={this.props.editorAccessibility.lintMessages}
-                  lineNumber={this.props.editorAccessibility.lineNumber}
-                  editorOptionsVisible={this.props.ide.editorOptionsVisible}
-                  showEditorOptions={this.props.showEditorOptions}
-                  closeEditorOptions={this.props.closeEditorOptions}
-                  showKeyboardShortcutModal={this.props.showKeyboardShortcutModal}
-                  setUnsavedChanges={this.props.setUnsavedChanges}
-                  isPlaying={this.props.ide.isPlaying}
-                  theme={this.props.preferences.theme}
-                  startRefreshSketch={this.props.startRefreshSketch}
-                  stopSketch={this.props.stopSketch}
-                  autorefresh={this.props.preferences.autorefresh}
-                  unsavedChanges={this.props.ide.unsavedChanges}
-                  projectSavedTime={this.props.ide.projectSavedTime}
-                  isExpanded={this.props.ide.sidebarIsExpanded}
-                  expandSidebar={this.props.expandSidebar}
-                  collapseSidebar={this.props.collapseSidebar}
-                  isUserOwner={this.isUserOwner()}
-                />
-                <Console
-                  consoleEvents={this.props.console}
-                  isPlaying={this.props.ide.isPlaying}
-                  isExpanded={this.props.ide.consoleIsExpanded}
-                  expandConsole={this.props.expandConsole}
-                  collapseConsole={this.props.collapseConsole}
-                  clearConsole={this.props.clearConsole}
-                />
+                <Editor provideController={(ctl) => { this.cmController = ctl; }} />
+                <Console />
               </SplitPane>
-              <div className="preview-frame-holder">
-                <div className="preview-frame-overlay" ref="overlay">
+              <section className="preview-frame-holder">
+                <header className="preview-frame__header">
+                  <h2 className="preview-frame__title">{this.props.t('Toolbar.Preview')}</h2>
+                </header>
+                <div className="preview-frame__content">
+                  <div
+                    className="preview-frame-overlay"
+                    ref={(element) => {
+                      this.overlay = element;
+                    }}
+                  >
+                  </div>
+                  <div>
+                    {((this.props.preferences.textOutput ||
+                      this.props.preferences.gridOutput ||
+                      this.props.preferences.soundOutput) &&
+                      this.props.ide.isPlaying) ||
+                      this.props.ide.isAccessibleOutputPlaying}
+                  </div>
+                  <PreviewFrame cmController={this.cmController} />
                 </div>
-                <div>
-                {(() => {
-                  if ((this.props.preferences.textOutput && this.props.ide.isPlaying) || this.props.ide.isTextOutputPlaying) {
-                    return (
-                      <TextOutput />
-                    );
-                  }
-                  return '';
-                })()}
-                </div>
-                <PreviewFrame
-                  htmlFile={this.props.htmlFile}
-                  files={this.props.files}
-                  content={this.props.selectedFile.content}
-                  isPlaying={this.props.ide.isPlaying}
-                  isTextOutputPlaying={this.props.ide.isTextOutputPlaying}
-                  textOutput={this.props.preferences.textOutput}
-                  setTextOutput={this.props.setTextOutput}
-                  dispatchConsoleEvent={this.props.dispatchConsoleEvent}
-                  autorefresh={this.props.preferences.autorefresh}
-                  previewIsRefreshing={this.props.ide.previewIsRefreshing}
-                  endSketchRefresh={this.props.endSketchRefresh}
-                  stopSketch={this.props.stopSketch}
-                  setBlobUrl={this.props.setBlobUrl}
-                  stopSketch={this.props.stopSketch}
-                  expandConsole={this.props.expandConsole}
-                />
-              </div>
+              </section>
             </SplitPane>
           </SplitPane>
-        </div>
-        {(() => {
-          if (this.props.ide.modalIsVisible) {
-            return (
-              <NewFileModal
-                canUploadMedia={this.props.user.authenticated}
-                closeModal={this.props.closeNewFileModal}
-                createFile={this.props.createFile}
-              />
-            );
-          }
-          return '';
-        })()}
-        {(() => {
-          if (this.props.ide.newFolderModalVisible) {
-            return (
-              <NewFolderModal
-                closeModal={this.props.closeNewFolderModal}
-                createFolder={this.props.createFolder}
-              />
-            );
-          }
-          return '';
-        })()}
-        {(() => { // eslint-disable-line
-          if (this.props.location.pathname.match(/sketches$/)) {
-            return (
-              <Overlay>
-                <SketchList
-                  username={this.props.params.username}
-                  previousPath={this.props.ide.previousPath}
-                />
-              </Overlay>
-            );
-          }
-        })()}
-        {(() => { // eslint-disable-line
-          if (this.props.location.pathname === '/about') {
-            return (
-              <Overlay>
-                <About previousPath={this.props.ide.previousPath} />
-              </Overlay>
-            );
-          }
-        })()}
-        {(() => { // eslint-disable-line
-          if (this.props.ide.shareModalVisible) {
-            return (
-              <Overlay>
-                <ShareModal
-                  projectId={this.props.project.id}
-                  closeShareModal={this.props.closeShareModal}
-                  ownerUsername={this.props.project.owner.username}
-                />
-              </Overlay>
-            );
-          }
-        })()}
-        {(() => { // eslint-disable-line
-          if (this.props.ide.keyboardShortcutVisible) {
-            return (
-              <Overlay>
-                <KeyboardShortcutModal
-                  closeModal={this.props.closeKeyboardShortcutModal}
-                />
-              </Overlay>
-            );
-          }
-        })()}
-        {(() => { // eslint-disable-line
-          if (this.props.ide.errorType) {
-            return (
-              <Overlay>
-                <ErrorModal
-                  type={this.props.ide.errorType}
-                  closeModal={this.props.hideErrorModal}
-                />
-              </Overlay>
-            );
-          }
-        })()}
+        </main>
+        {this.props.ide.modalIsVisible && <NewFileModal />}
+        {this.props.ide.newFolderModalVisible && (
+          <NewFolderModal
+            closeModal={this.props.closeNewFolderModal}
+            createFolder={this.props.createFolder}
+          />
+        )}
+        {this.props.ide.uploadFileModalVisible && (
+          <UploadFileModal closeModal={this.props.closeUploadFileModal} />
+        )}
+        {this.props.location.pathname === '/about' && (
+          <Overlay
+            title={this.props.t('About.Title')}
+            previousPath={this.props.ide.previousPath}
+            ariaLabel={this.props.t('About.Title')}
+          >
+            <About previousPath={this.props.ide.previousPath} />
+          </Overlay>
+        )}
+        {this.props.location.pathname === '/feedback' && (
+          <Overlay
+            title={this.props.t('IDEView.SubmitFeedback')}
+            previousPath={this.props.ide.previousPath}
+            ariaLabel={this.props.t('IDEView.SubmitFeedbackARIA')}
+          >
+            <Feedback previousPath={this.props.ide.previousPath} />
+          </Overlay>
+        )}
+        {this.props.location.pathname.match(/add-to-collection$/) && (
+          <Overlay
+            ariaLabel={this.props.t('IDEView.AddCollectionARIA')}
+            title={this.props.t('IDEView.AddCollectionTitle')}
+            previousPath={this.props.ide.previousPath}
+            actions={<CollectionSearchbar />}
+            isFixedHeight
+          >
+            <AddToCollectionList
+              projectId={this.props.params.project_id}
+              username={this.props.params.username}
+              user={this.props.user}
+            />
+          </Overlay>
+        )}
+        {this.props.ide.shareModalVisible && (
+          <Overlay
+            title={this.props.t('IDEView.ShareTitle')}
+            ariaLabel={this.props.t('IDEView.ShareARIA')}
+            closeOverlay={this.props.closeShareModal}
+          >
+            <ShareModal
+              projectId={this.props.ide.shareModalProjectId}
+              projectName={this.props.ide.shareModalProjectName}
+              ownerUsername={this.props.ide.shareModalProjectUsername}
+            />
+          </Overlay>
+        )}
+        {this.props.ide.keyboardShortcutVisible && (
+          <Overlay
+            title={this.props.t('KeyboardShortcuts.Title')}
+            ariaLabel={this.props.t('KeyboardShortcuts.Title')}
+            closeOverlay={this.props.closeKeyboardShortcutModal}
+          >
+            <KeyboardShortcutModal />
+          </Overlay>
+        )}
+        {this.props.ide.errorType && (
+          <Overlay
+            title={this.props.t('Common.Error')}
+            ariaLabel={this.props.t('Common.ErrorARIA')}
+            closeOverlay={this.props.hideErrorModal}
+          >
+            <ErrorModal
+              type={this.props.ide.errorType}
+              closeModal={this.props.hideErrorModal}
+            />
+          </Overlay>
+        )}
       </div>
-
     );
   }
 }
@@ -448,148 +456,135 @@ IDEView.propTypes = {
     project_id: PropTypes.string,
     username: PropTypes.string,
     reset_password_token: PropTypes.string,
-  }),
+  }).isRequired,
   location: PropTypes.shape({
-    pathname: PropTypes.string
-  }),
+    pathname: PropTypes.string,
+  }).isRequired,
   getProject: PropTypes.func.isRequired,
   user: PropTypes.shape({
     authenticated: PropTypes.bool.isRequired,
     id: PropTypes.string,
-    username: PropTypes.string
+    username: PropTypes.string,
   }).isRequired,
-  newProject: PropTypes.func.isRequired,
   saveProject: PropTypes.func.isRequired,
   ide: PropTypes.shape({
+    errorType: PropTypes.string,
+    keyboardShortcutVisible: PropTypes.bool.isRequired,
+    shareModalVisible: PropTypes.bool.isRequired,
+    shareModalProjectId: PropTypes.string.isRequired,
+    shareModalProjectName: PropTypes.string.isRequired,
+    shareModalProjectUsername: PropTypes.string.isRequired,
+    previousPath: PropTypes.string.isRequired,
+    previewIsRefreshing: PropTypes.bool.isRequired,
     isPlaying: PropTypes.bool.isRequired,
-    isTextOutputPlaying: PropTypes.bool.isRequired,
-    consoleEvent: PropTypes.array,
+    isAccessibleOutputPlaying: PropTypes.bool.isRequired,
+    projectOptionsVisible: PropTypes.bool.isRequired,
+    preferencesIsVisible: PropTypes.bool.isRequired,
     modalIsVisible: PropTypes.bool.isRequired,
+    uploadFileModalVisible: PropTypes.bool.isRequired,
+    newFolderModalVisible: PropTypes.bool.isRequired,
+    justOpenedProject: PropTypes.bool.isRequired,
     sidebarIsExpanded: PropTypes.bool.isRequired,
     consoleIsExpanded: PropTypes.bool.isRequired,
-    preferencesIsVisible: PropTypes.bool.isRequired,
-    projectOptionsVisible: PropTypes.bool.isRequired,
-    newFolderModalVisible: PropTypes.bool.isRequired,
-    shareModalVisible: PropTypes.bool.isRequired,
-    editorOptionsVisible: PropTypes.bool.isRequired,
-    keyboardShortcutVisible: PropTypes.bool.isRequired,
     unsavedChanges: PropTypes.bool.isRequired,
-    infiniteLoop: PropTypes.bool.isRequired,
-    previewIsRefreshing: PropTypes.bool.isRequired,
-    infiniteLoopMessage: PropTypes.string.isRequired,
-    projectSavedTime: PropTypes.string.isRequired,
-    previousPath: PropTypes.string.isRequired,
-    justOpenedProject: PropTypes.bool.isRequired,
-    errorType: PropTypes.string
   }).isRequired,
-  startSketch: PropTypes.func.isRequired,
   stopSketch: PropTypes.func.isRequired,
-  startTextOutput: PropTypes.func.isRequired,
-  stopTextOutput: PropTypes.func.isRequired,
-  detectInfiniteLoops: PropTypes.func.isRequired,
-  resetInfiniteLoops: PropTypes.func.isRequired,
   project: PropTypes.shape({
     id: PropTypes.string,
     name: PropTypes.string.isRequired,
     owner: PropTypes.shape({
       username: PropTypes.string,
-      id: PropTypes.string
-    })
+      id: PropTypes.string,
+    }),
+    updatedAt: PropTypes.string,
   }).isRequired,
-  setProjectName: PropTypes.func.isRequired,
-  openPreferences: PropTypes.func.isRequired,
   editorAccessibility: PropTypes.shape({
-    lintMessages: PropTypes.array.isRequired,
-    lineNumber: PropTypes.string.isRequired
+    lintMessages: PropTypes.array.isRequired, // eslint-disable-line
   }).isRequired,
-  updateLintMessage: PropTypes.func.isRequired,
-  clearLintMessage: PropTypes.func.isRequired,
-  updateLineNumber: PropTypes.func.isRequired,
   preferences: PropTypes.shape({
-    fontSize: PropTypes.number.isRequired,
-    indentationAmount: PropTypes.number.isRequired,
-    isTabIndent: PropTypes.bool.isRequired,
     autosave: PropTypes.bool.isRequired,
+    fontSize: PropTypes.number.isRequired,
+    linewrap: PropTypes.bool.isRequired,
+    lineNumbers: PropTypes.bool.isRequired,
     lintWarning: PropTypes.bool.isRequired,
-    textOutput: PropTypes.number.isRequired,
+    textOutput: PropTypes.bool.isRequired,
+    gridOutput: PropTypes.bool.isRequired,
+    soundOutput: PropTypes.bool.isRequired,
     theme: PropTypes.string.isRequired,
-    autorefresh: PropTypes.bool.isRequired
+    autorefresh: PropTypes.bool.isRequired,
+    language: PropTypes.string.isRequired,
+    autocloseBracketsQuotes: PropTypes.bool.isRequired
   }).isRequired,
   closePreferences: PropTypes.func.isRequired,
+  setAutocloseBracketsQuotes: PropTypes.func.isRequired,
   setFontSize: PropTypes.func.isRequired,
-  setIndentation: PropTypes.func.isRequired,
-  indentWithTab: PropTypes.func.isRequired,
-  indentWithSpace: PropTypes.func.isRequired,
   setAutosave: PropTypes.func.isRequired,
+  setLineNumbers: PropTypes.func.isRequired,
+  setLinewrap: PropTypes.func.isRequired,
   setLintWarning: PropTypes.func.isRequired,
   setTextOutput: PropTypes.func.isRequired,
-  files: PropTypes.array.isRequired,
-  updateFileContent: PropTypes.func.isRequired,
+  setGridOutput: PropTypes.func.isRequired,
+  setSoundOutput: PropTypes.func.isRequired,
+  setAllAccessibleOutput: PropTypes.func.isRequired,
+  files: PropTypes.arrayOf(PropTypes.shape({
+    id: PropTypes.string.isRequired,
+    name: PropTypes.string.isRequired,
+    content: PropTypes.string.isRequired,
+  })).isRequired,
   selectedFile: PropTypes.shape({
     id: PropTypes.string.isRequired,
-    content: PropTypes.string.isRequired
-  }),
+    content: PropTypes.string.isRequired,
+    name: PropTypes.string.isRequired,
+  }).isRequired,
   setSelectedFile: PropTypes.func.isRequired,
-  htmlFile: PropTypes.object.isRequired,
-  dispatchConsoleEvent: PropTypes.func.isRequired,
+  htmlFile: PropTypes.shape({
+    id: PropTypes.string.isRequired,
+    name: PropTypes.string.isRequired,
+    content: PropTypes.string.isRequired,
+  }).isRequired,
   newFile: PropTypes.func.isRequired,
-  closeNewFileModal: PropTypes.func.isRequired,
   expandSidebar: PropTypes.func.isRequired,
   collapseSidebar: PropTypes.func.isRequired,
-  exportProjectAsZip: PropTypes.func.isRequired,
   cloneProject: PropTypes.func.isRequired,
   expandConsole: PropTypes.func.isRequired,
   collapseConsole: PropTypes.func.isRequired,
-  showFileOptions: PropTypes.func.isRequired,
-  hideFileOptions: PropTypes.func.isRequired,
   deleteFile: PropTypes.func.isRequired,
-  showEditFileName: PropTypes.func.isRequired,
-  hideEditFileName: PropTypes.func.isRequired,
   updateFileName: PropTypes.func.isRequired,
-  showEditProjectName: PropTypes.func.isRequired,
-  hideEditProjectName: PropTypes.func.isRequired,
-  logoutUser: PropTypes.func.isRequired,
   openProjectOptions: PropTypes.func.isRequired,
   closeProjectOptions: PropTypes.func.isRequired,
   newFolder: PropTypes.func.isRequired,
   closeNewFolderModal: PropTypes.func.isRequired,
+  closeNewFileModal: PropTypes.func.isRequired,
   createFolder: PropTypes.func.isRequired,
-  createFile: PropTypes.func.isRequired,
-  showShareModal: PropTypes.func.isRequired,
   closeShareModal: PropTypes.func.isRequired,
-  showEditorOptions: PropTypes.func.isRequired,
-  closeEditorOptions: PropTypes.func.isRequired,
-  showKeyboardShortcutModal: PropTypes.func.isRequired,
   closeKeyboardShortcutModal: PropTypes.func.isRequired,
   toast: PropTypes.shape({
-    isVisible: PropTypes.bool.isRequired
+    isVisible: PropTypes.bool.isRequired,
   }).isRequired,
-  showToast: PropTypes.func.isRequired,
-  setToastText: PropTypes.func.isRequired,
   autosaveProject: PropTypes.func.isRequired,
   router: PropTypes.shape({
-    setRouteLeaveHook: PropTypes.func
+    setRouteLeaveHook: PropTypes.func,
   }).isRequired,
-  route: PropTypes.object.isRequired,
-  setUnsavedChanges: PropTypes.func.isRequired,
+  route: PropTypes.oneOfType([PropTypes.object, PropTypes.element]).isRequired,
   setTheme: PropTypes.func.isRequired,
-  setAutorefresh: PropTypes.func.isRequired,
-  startSketchAndRefresh: PropTypes.func.isRequired,
-  endSketchRefresh: PropTypes.func.isRequired,
-  startRefreshSketch: PropTypes.func.isRequired,
-  setBlobUrl: PropTypes.func.isRequired,
   setPreviousPath: PropTypes.func.isRequired,
-  resetProject: PropTypes.func.isRequired,
-  console: PropTypes.array.isRequired,
-  clearConsole: PropTypes.func.isRequired,
   showErrorModal: PropTypes.func.isRequired,
-  hideErrorModal: PropTypes.func.isRequired
+  hideErrorModal: PropTypes.func.isRequired,
+  clearPersistedState: PropTypes.func.isRequired,
+  startSketch: PropTypes.func.isRequired,
+  openUploadFileModal: PropTypes.func.isRequired,
+  closeUploadFileModal: PropTypes.func.isRequired,
+  t: PropTypes.func.isRequired,
+  isUserOwner: PropTypes.bool.isRequired
 };
 
 function mapStateToProps(state) {
   return {
     files: state.files,
-    selectedFile: state.files.find(file => file.isSelectedFile),
+    selectedFile:
+      state.files.find(file => file.isSelectedFile) ||
+      state.files.find(file => file.name === 'sketch.js') ||
+      state.files.find(file => file.name !== 'root'),
     htmlFile: getHTMLFile(state.files),
     ide: state.ide,
     preferences: state.preferences,
@@ -597,21 +592,27 @@ function mapStateToProps(state) {
     user: state.user,
     project: state.project,
     toast: state.toast,
-    console: state.console
+    console: state.console,
+    isUserOwner: getIsUserOwner(state)
   };
 }
 
 function mapDispatchToProps(dispatch) {
-  return bindActionCreators(Object.assign({},
-    EditorAccessibilityActions,
-    FileActions,
-    ProjectActions,
-    IDEActions,
-    PreferencesActions,
-    UserActions,
-    ToastActions,
-    ConsoleActions),
-  dispatch);
+  return bindActionCreators(
+    Object.assign(
+      {},
+      EditorAccessibilityActions,
+      FileActions,
+      ProjectActions,
+      IDEActions,
+      PreferencesActions,
+      UserActions,
+      ToastActions,
+      ConsoleActions
+    ),
+    dispatch
+  );
 }
 
-export default withRouter(connect(mapStateToProps, mapDispatchToProps)(IDEView));
+
+export default withTranslation()(withRouter(connect(mapStateToProps, mapDispatchToProps)(IDEView)));
