@@ -1,22 +1,23 @@
-import * as ActionTypes from '../../../constants';
 import objectID from 'bson-objectid';
+import * as ActionTypes from '../../../constants';
 
-const defaultSketch = `function setup() { 
+const defaultSketch = `function setup() {
   createCanvas(400, 400);
-} 
+}
 
-function draw() { 
+function draw() {
   background(220);
 }`;
 
 const defaultHTML =
 `<!DOCTYPE html>
-<html>
+<html lang="en">
   <head>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/p5.js/0.5.5/p5.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/p5.js/0.5.5/addons/p5.dom.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/p5.js/0.5.5/addons/p5.sound.min.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/p5.js/1.1.9/p5.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/p5.js/1.1.9/addons/p5.sound.min.js"></script>
     <link rel="stylesheet" type="text/css" href="style.css">
+    <meta charset="utf-8" />
+
   </head>
   <body>
     <script src="sketch.js"></script>
@@ -28,6 +29,9 @@ const defaultCSS =
 `html, body {
   margin: 0;
   padding: 0;
+}
+canvas {
+  display: block;
 }
 `;
 
@@ -41,8 +45,9 @@ const initialState = () => {
       name: 'root',
       id: r,
       _id: r,
-      children: [a, b, c],
-      fileType: 'folder'
+      children: [b, a, c],
+      fileType: 'folder',
+      content: ''
     },
     {
       name: 'sketch.js',
@@ -73,9 +78,9 @@ const initialState = () => {
 
 function getAllDescendantIds(state, nodeId) {
   return state.find(file => file.id === nodeId).children
-  .reduce((acc, childId) => (
-    [...acc, childId, ...getAllDescendantIds(state, childId)]
-  ), []);
+    .reduce((acc, childId) => (
+      [...acc, childId, ...getAllDescendantIds(state, childId)]
+    ), []);
 }
 
 function deleteChild(state, parentId, id) {
@@ -92,7 +97,7 @@ function deleteChild(state, parentId, id) {
 
 function deleteMany(state, ids) {
   const newState = [...state];
-  ids.forEach(id => {
+  ids.forEach((id) => {
     let fileIndex;
     newState.find((file, index) => {
       if (file.id === id) {
@@ -105,22 +110,48 @@ function deleteMany(state, ids) {
   return newState;
 }
 
+function sortedChildrenId(state, children) {
+  const childrenArray = state.filter(file => children.includes(file.id));
+  childrenArray.sort((a, b) => (a.name > b.name ? 1 : -1));
+  return childrenArray.map(child => child.id);
+}
+
+function updateParent(state, action) {
+  return state.map((file) => {
+    if (file.id === action.parentId) {
+      const newFile = Object.assign({}, file);
+      newFile.children = [...newFile.children, action.id];
+      return newFile;
+    }
+    return file;
+  });
+}
+
+function renameFile(state, action) {
+  return state.map((file) => {
+    if (file.id !== action.id) {
+      return file;
+    }
+    return Object.assign({}, file, { name: action.name });
+  });
+}
+
 const files = (state, action) => {
   if (state === undefined) {
     state = initialState(); // eslint-disable-line
   }
   switch (action.type) {
     case ActionTypes.UPDATE_FILE_CONTENT:
-      return state.map(file => {
-        if (file.name !== action.name) {
+      return state.map((file) => {
+        if (file.id !== action.id) {
           return file;
         }
 
         return Object.assign({}, file, { content: action.content });
       });
     case ActionTypes.SET_BLOB_URL:
-      return state.map(file => {
-        if (file.name !== action.name) {
+      return state.map((file) => {
+        if (file.id !== action.id) {
           return file;
         }
         return Object.assign({}, file, { blobURL: action.blobURL });
@@ -132,106 +163,80 @@ const files = (state, action) => {
     case ActionTypes.RESET_PROJECT:
       return initialState();
     case ActionTypes.CREATE_FILE: // eslint-disable-line
-      {
-        const newState = state.map((file) => {
-          if (file.id === action.parentId) {
-            const newFile = Object.assign({}, file);
-            newFile.children = [...newFile.children, action.id];
-            return newFile;
-          }
-          return file;
-        });
-        return [...newState,
-          { name: action.name,
-            id: action.id,
-            _id: action._id,
-            content: action.content,
-            url: action.url,
-            children: action.children,
-            fileType: action.fileType || 'file' }];
-      }
-    case ActionTypes.SHOW_FILE_OPTIONS:
-      return state.map(file => {
-        if (file.id !== action.id) {
-          return file;
+    {
+      const newState = [
+        ...updateParent(state, action),
+        {
+          name: action.name,
+          id: action.id,
+          _id: action._id,
+          content: action.content,
+          url: action.url,
+          children: action.children,
+          fileType: action.fileType || 'file'
+        }];
+      return newState.map((file) => {
+        if (file.id === action.parentId) {
+          file.children = sortedChildrenId(newState, file.children);
         }
-
-        return Object.assign({}, file, { isOptionsOpen: true });
+        return file;
       });
-    case ActionTypes.HIDE_FILE_OPTIONS:
-      return state.map(file => {
-        if (file.id !== action.id) {
-          return file;
-        }
-
-        return Object.assign({}, file, { isOptionsOpen: false });
-      });
+    }
     case ActionTypes.UPDATE_FILE_NAME:
-      return state.map(file => {
-        if (file.id !== action.id) {
-          return file;
+    {
+      const newState = renameFile(state, action);
+      return newState.map((file) => {
+        if (file.children.includes(action.id)) {
+          file.children = sortedChildrenId(newState, file.children);
         }
-
-        return Object.assign({}, file, { name: action.name });
+        return file;
       });
+    }
     case ActionTypes.DELETE_FILE:
-      {
-        const newState = deleteMany(state, [action.id, ...getAllDescendantIds(state, action.id)]);
-        return deleteChild(newState, action.parentId, action.id);
-        // const newState = state.map((file) => {
-        //   if (file.id === action.parentId) {
-        //     const newChildren = file.children.filter(child => child !== action.id);
-        //     return { ...file, children: newChildren };
-        //   }
-        //   return file;
-        // });
-        // return newState.filter(file => file.id !== action.id);
-      }
-    case ActionTypes.SHOW_EDIT_FILE_NAME:
-      return state.map(file => {
-        if (file.id !== action.id) {
-          return file;
-        }
-
-        return Object.assign({}, file, { isEditingName: true });
-      });
-    case ActionTypes.HIDE_EDIT_FILE_NAME:
-      return state.map(file => {
-        if (file.id !== action.id) {
-          return file;
-        }
-
-        return Object.assign({}, file, { isEditingName: false });
-      });
+    {
+      const newState = deleteMany(state, [action.id, ...getAllDescendantIds(state, action.id)]);
+      return deleteChild(newState, action.parentId, action.id);
+      // const newState = state.map((file) => {
+      //   if (file.id === action.parentId) {
+      //     const newChildren = file.children.filter(child => child !== action.id);
+      //     return { ...file, children: newChildren };
+      //   }
+      //   return file;
+      // });
+      // return newState.filter(file => file.id !== action.id);
+    }
     case ActionTypes.SET_SELECTED_FILE:
-      return state.map(file => {
+      return state.map((file) => {
         if (file.id === action.selectedFile) {
           return Object.assign({}, file, { isSelectedFile: true });
         }
         return Object.assign({}, file, { isSelectedFile: false });
       });
     case ActionTypes.SHOW_FOLDER_CHILDREN:
-      return state.map(file => {
+      return state.map((file) => {
         if (file.id === action.id) {
           return Object.assign({}, file, { isFolderClosed: false });
         }
         return file;
       });
     case ActionTypes.HIDE_FOLDER_CHILDREN:
-      return state.map(file => {
+      return state.map((file) => {
         if (file.id === action.id) {
           return Object.assign({}, file, { isFolderClosed: true });
         }
         return file;
       });
     default:
-      return state;
+      return state.map((file) => {
+        file.children = sortedChildrenId(state, file.children);
+        return file;
+      });
   }
 };
 
-export const getHTMLFile = (state) => state.filter(file => file.name.match(/.*\.html$/i))[0];
-export const getJSFiles = (state) => state.filter(file => file.name.match(/.*\.js$/i));
-export const getCSSFiles = (state) => state.filter(file => file.name.match(/.*\.css$/i));
-export const getLinkedFiles = (state) => state.filter(file => file.url);
+export const getHTMLFile = state => state.filter(file => file.name.match(/.*\.html$/i))[0];
+export const getJSFiles = state => state.filter(file => file.name.match(/.*\.js$/i));
+export const getCSSFiles = state => state.filter(file => file.name.match(/.*\.css$/i));
+export const getLinkedFiles = state => state.filter(file => file.url);
 
 export default files;
