@@ -2,14 +2,18 @@ import * as ActionTypes from '../../../constants';
 
 const initialState = {
   isPlaying: false,
-  isTextOutputPlaying: false,
+  isAccessibleOutputPlaying: false,
   modalIsVisible: false,
   sidebarIsExpanded: false,
   consoleIsExpanded: true,
   preferencesIsVisible: false,
   projectOptionsVisible: false,
   newFolderModalVisible: false,
+  uploadFileModalVisible: false,
   shareModalVisible: false,
+  shareModalProjectId: 'abcd',
+  shareModalProjectName: 'My Cute Sketch',
+  shareModalProjectUsername: 'p5_user',
   editorOptionsVisible: false,
   keyboardShortcutVisible: false,
   unsavedChanges: false,
@@ -17,9 +21,10 @@ const initialState = {
   previewIsRefreshing: false,
   infiniteLoopMessage: '',
   justOpenedProject: false,
-  projectSavedTime: '',
   previousPath: '/',
-  errorType: undefined
+  errorType: undefined,
+  runtimeErrorWarningVisible: true,
+  parentId: undefined
 };
 
 const ide = (state = initialState, action) => {
@@ -28,14 +33,18 @@ const ide = (state = initialState, action) => {
       return Object.assign({}, state, { isPlaying: true });
     case ActionTypes.STOP_SKETCH:
       return Object.assign({}, state, { isPlaying: false });
-    case ActionTypes.START_TEXT_OUTPUT:
-      return Object.assign({}, state, { isTextOutputPlaying: true });
-    case ActionTypes.STOP_TEXT_OUTPUT:
-      return Object.assign({}, state, { isTextOutputPlaying: false });
+    case ActionTypes.START_ACCESSIBLE_OUTPUT:
+      return Object.assign({}, state, { isAccessibleOutputPlaying: true });
+    case ActionTypes.STOP_ACCESSIBLE_OUTPUT:
+      return Object.assign({}, state, { isAccessibleOutputPlaying: false });
     case ActionTypes.CONSOLE_EVENT:
       return Object.assign({}, state, { consoleEvent: action.event });
     case ActionTypes.SHOW_MODAL:
-      return Object.assign({}, state, { modalIsVisible: true });
+      return Object.assign({}, state, {
+        modalIsVisible: true,
+        parentId: action.parentId,
+        newFolderModalVisible: false
+      });
     case ActionTypes.HIDE_MODAL:
       return Object.assign({}, state, { modalIsVisible: false });
     case ActionTypes.COLLAPSE_SIDEBAR:
@@ -57,11 +66,20 @@ const ide = (state = initialState, action) => {
     case ActionTypes.CLOSE_PROJECT_OPTIONS:
       return Object.assign({}, state, { projectOptionsVisible: false });
     case ActionTypes.SHOW_NEW_FOLDER_MODAL:
-      return Object.assign({}, state, { newFolderModalVisible: true });
+      return Object.assign({}, state, {
+        newFolderModalVisible: true,
+        parentId: action.parentId,
+        modalIsVisible: false
+      });
     case ActionTypes.CLOSE_NEW_FOLDER_MODAL:
       return Object.assign({}, state, { newFolderModalVisible: false });
     case ActionTypes.SHOW_SHARE_MODAL:
-      return Object.assign({}, state, { shareModalVisible: true });
+      return Object.assign({}, state, {
+        shareModalVisible: true,
+        shareModalProjectId: action.payload.shareModalProjectId,
+        shareModalProjectName: action.payload.shareModalProjectName,
+        shareModalProjectUsername: action.payload.shareModalProjectUsername,
+      });
     case ActionTypes.CLOSE_SHARE_MODAL:
       return Object.assign({}, state, { shareModalVisible: false });
     case ActionTypes.SHOW_EDITOR_OPTIONS:
@@ -86,16 +104,20 @@ const ide = (state = initialState, action) => {
       return Object.assign({}, state, { justOpenedProject: true });
     case ActionTypes.RESET_JUST_OPENED_PROJECT:
       return Object.assign({}, state, { justOpenedProject: false });
-    case ActionTypes.SET_PROJECT_SAVED_TIME:
-      return Object.assign({}, state, { projectSavedTime: action.value });
-    case ActionTypes.RESET_PROJECT_SAVED_TIME:
-      return Object.assign({}, state, { projectSavedTime: '' });
     case ActionTypes.SET_PREVIOUS_PATH:
       return Object.assign({}, state, { previousPath: action.path });
     case ActionTypes.SHOW_ERROR_MODAL:
       return Object.assign({}, state, { errorType: action.modalType });
     case ActionTypes.HIDE_ERROR_MODAL:
       return Object.assign({}, state, { errorType: undefined });
+    case ActionTypes.HIDE_RUNTIME_ERROR_WARNING:
+      return Object.assign({}, state, { runtimeErrorWarningVisible: false });
+    case ActionTypes.SHOW_RUNTIME_ERROR_WARNING:
+      return Object.assign({}, state, { runtimeErrorWarningVisible: true });
+    case ActionTypes.OPEN_UPLOAD_FILE_MODAL:
+      return Object.assign({}, state, { uploadFileModalVisible: true, parentId: action.parentId });
+    case ActionTypes.CLOSE_UPLOAD_FILE_MODAL:
+      return Object.assign({}, state, { uploadFileModalVisible: false });
     default:
       return state;
   }
