@@ -1,11 +1,13 @@
 import * as ActionTypes from '../../../constants';
-import generate from 'project-name-generator';
+import { generateProjectName } from '../../../utils/generateRandomName';
 
 const initialState = () => {
-  const generatedString = generate({ words: 2 }).spaced;
+  const generatedString = generateProjectName();
   const generatedName = generatedString.charAt(0).toUpperCase() + generatedString.slice(1);
   return {
-    name: generatedName
+    name: generatedName,
+    updatedAt: '',
+    isSaving: false
   };
 };
 
@@ -21,14 +23,16 @@ const project = (state, action) => {
         id: action.project.id,
         name: action.project.name,
         updatedAt: action.project.updatedAt,
-        owner: action.owner
+        owner: action.owner,
+        isSaving: false
       };
     case ActionTypes.SET_PROJECT:
       return {
         id: action.project.id,
         name: action.project.name,
         updatedAt: action.project.updatedAt,
-        owner: action.owner
+        owner: action.owner,
+        isSaving: false
       };
     case ActionTypes.RESET_PROJECT:
       return initialState();
@@ -36,6 +40,12 @@ const project = (state, action) => {
       return Object.assign({}, state, { isEditingName: true });
     case ActionTypes.HIDE_EDIT_PROJECT_NAME:
       return Object.assign({}, state, { isEditingName: false });
+    case ActionTypes.SET_PROJECT_SAVED_TIME:
+      return Object.assign({}, state, { updatedAt: action.value });
+    case ActionTypes.START_SAVING_PROJECT:
+      return Object.assign({}, state, { isSaving: true });
+    case ActionTypes.END_SAVING_PROJECT:
+      return Object.assign({}, state, { isSaving: false });
     default:
       return state;
   }
