@@ -1,10 +1,9 @@
-import * as ActionTypes from '../../constants';
 import { browserHistory } from 'react-router';
-import axios from 'axios';
+import * as ActionTypes from '../../constants';
+import apiClient from '../../utils/apiClient';
 import { showErrorModal, justOpenedProject } from '../IDE/actions/ide';
-
-
-const ROOT_URL = location.href.indexOf('localhost') > 0 ? 'http://localhost:8000/api' : '/api';
+import { setLanguage } from '../IDE/actions/preferences';
+import { showToast, setToastText } from '../IDE/actions/toast';
 
 export function authError(error) {
   return {
@@ -15,20 +14,24 @@ export function authError(error) {
 
 export function signUpUser(previousPath, formValues) {
   return (dispatch) => {
-    axios.post(`${ROOT_URL}/signup`, formValues, { withCredentials: true })
-      .then(response => {
-        dispatch({ type: ActionTypes.AUTH_USER,
-                    user: response.data
+    apiClient.post('/signup', formValues)
+      .then((response) => {
+        dispatch({
+          type: ActionTypes.AUTH_USER,
+          user: response.data
         });
         dispatch(justOpenedProject());
         browserHistory.push(previousPath);
       })
-      .catch(response => dispatch(authError(response.data.error)));
+      .catch((error) => {
+        const { response } = error;
+        dispatch(authError(response.data.error));
+      });
   };
 }
 
 export function loginUser(formValues) {
-  return axios.post(`${ROOT_URL}/login`, formValues, { withCredentials: true });
+  return apiClient.post('/login', formValues);
 }
 
 export function loginUserSuccess(user) {
@@ -48,28 +51,7 @@ export function loginUserFailure(error) {
 export function validateAndLoginUser(previousPath, formProps, dispatch) {
   return new Promise((resolve, reject) => {
     loginUser(formProps)
-      .then(response => {
-        dispatch({ type: ActionTypes.AUTH_USER,
-                  user: response.data
-        });
-        dispatch({
-          type: ActionTypes.SET_PREFERENCES,
-          preferences: response.data.preferences
-        });
-        dispatch(justOpenedProject());
-        browserHistory.push(previousPath);
-        resolve();
-      })
-      .catch(response => {
-        reject({ password: response.data.message, _error: 'Login failed!' });
-      });
-  });
-}
-
-export function getUser() {
-  return (dispatch) => {
-    axios.get(`${ROOT_URL}/session`, { withCredentials: true })
-      .then(response => {
+      .then((response) => {
         dispatch({
           type: ActionTypes.AUTH_USER,
           user: response.data
@@ -78,23 +60,48 @@ export function getUser() {
           type: ActionTypes.SET_PREFERENCES,
           preferences: response.data.preferences
         });
+        setLanguage(response.data.preferences.language, { persistPreference: false });
+        dispatch(justOpenedProject());
+        browserHistory.push(previousPath);
+        resolve();
       })
-      .catch(response => {
-        dispatch(authError(response.data.error));
+      .catch(error =>
+        reject({ password: error.response.data.message, _error: 'Login failed!' })); // eslint-disable-line
+  });
+}
+
+export function getUser() {
+  return (dispatch) => {
+    apiClient.get('/session')
+      .then((response) => {
+        dispatch({
+          type: ActionTypes.AUTH_USER,
+          user: response.data
+        });
+        dispatch({
+          type: ActionTypes.SET_PREFERENCES,
+          preferences: response.data.preferences
+        });
+        setLanguage(response.data.preferences.language, { persistPreference: false });
+      }).catch((error) => {
+        const { response } = error;
+        const message = response.message || response.data.error;
+        dispatch(authError(message));
       });
   };
 }
 
 export function validateSession() {
   return (dispatch, getState) => {
-    axios.get(`${ROOT_URL}/session`, { withCredentials: true })
-      .then(response => {
+    apiClient.get('/session')
+      .then((response) => {
         const state = getState();
         if (state.user.username !== response.data.username) {
           dispatch(showErrorModal('staleSession'));
         }
       })
-      .catch(response => {
+      .catch((error) => {
+        const { response } = error;
         if (response.status === 404) {
           dispatch(showErrorModal('staleSession'));
         }
@@ -104,13 +111,16 @@ export function validateSession() {
 
 export function logoutUser() {
   return (dispatch) => {
-    axios.get(`${ROOT_URL}/logout`, { withCredentials: true })
+    apiClient.get('/logout')
       .then(() => {
         dispatch({
           type: ActionTypes.UNAUTH_USER
         });
       })
-      .catch(response => dispatch(authError(response.data.error)));
+      .catch((error) => {
+        const { response } = error;
+        dispatch(authError(response.data.error));
+      });
   };
 }
 
@@ -119,16 +129,60 @@ export function initiateResetPassword(formValues) {
     dispatch({
       type: ActionTypes.RESET_PASSWORD_INITIATE
     });
-    axios.post(`${ROOT_URL}/reset-password`, formValues, { withCredentials: true })
+    apiClient.post('/reset-password', formValues)
       .then(() => {
         // do nothing
       })
-      .catch(response => dispatch({
-        type: ActionTypes.ERROR,
-        message: response.data
-      }));
+      .catch((error) => {
+        const { response } = error;
+        dispatch({
+          type: ActionTypes.ERROR,
+          message: response.data
+        });
+      });
   };
 }
+
+export function initiateVerification() {
+  return (dispatch) => {
+    dispatch({
+      type: ActionTypes.EMAIL_VERIFICATION_INITIATE
+    });
+    apiClient.post('/verify/send', {})
+      .then(() => {
+        // do nothing
+      })
+      .catch((error) => {
+        const { response } = error;
+        dispatch({
+          type: ActionTypes.ERROR,
+          message: response.data
+        });
+      });
+  };
+}
+
+export function verifyEmailConfirmation(token) {
+  return (dispatch) => {
+    dispatch({
+      type: ActionTypes.EMAIL_VERIFICATION_VERIFY,
+      state: 'checking',
+    });
+    return apiClient.get(`/verify?t=${token}`, {})
+      .then(response => dispatch({
+        type: ActionTypes.EMAIL_VERIFICATION_VERIFIED,
+        message: response.data,
+      }))
+      .catch((error) => {
+        const { response } = error;
+        dispatch({
+          type: ActionTypes.EMAIL_VERIFICATION_INVALID,
+          message: response.data
+        });
+      });
+  };
+}
+
 
 export function resetPasswordReset() {
   return {
@@ -138,7 +192,7 @@ export function resetPasswordReset() {
 
 export function validateResetPasswordToken(token) {
   return (dispatch) => {
-    axios.get(`${ROOT_URL}/reset-password/${token}`)
+    apiClient.get(`/reset-password/${token}`)
       .then(() => {
         // do nothing if the token is valid
       })
@@ -150,7 +204,7 @@ export function validateResetPasswordToken(token) {
 
 export function updatePassword(token, formValues) {
   return (dispatch) => {
-    axios.post(`${ROOT_URL}/reset-password/${token}`, formValues)
+    apiClient.post(`/reset-password/${token}`, formValues)
       .then((response) => {
         dispatch(loginUserSuccess(response.data));
         browserHistory.push('/');
@@ -158,5 +212,78 @@ export function updatePassword(token, formValues) {
       .catch(() => dispatch({
         type: ActionTypes.INVALID_RESET_PASSWORD_TOKEN
       }));
+  };
+}
+
+export function updateSettingsSuccess(user) {
+  return {
+    type: ActionTypes.SETTINGS_UPDATED,
+    user
+  };
+}
+
+export function updateSettings(formValues) {
+  return dispatch =>
+    apiClient.put('/account', formValues)
+      .then((response) => {
+        dispatch(updateSettingsSuccess(response.data));
+        browserHistory.push('/');
+        dispatch(showToast(5500));
+        dispatch(setToastText('Toast.SettingsSaved'));
+      })
+      .catch((error) => {
+        const { response } = error;
+        Promise.reject(new Error(response.data.error));
+      });
+}
+
+export function createApiKeySuccess(user) {
+  return {
+    type: ActionTypes.API_KEY_CREATED,
+    user
+  };
+}
+
+export function createApiKey(label) {
+  return dispatch =>
+    apiClient.post('/account/api-keys', { label })
+      .then((response) => {
+        dispatch(createApiKeySuccess(response.data));
+      })
+      .catch((error) => {
+        const { response } = error;
+        Promise.reject(new Error(response.data.error));
+      });
+}
+
+export function removeApiKey(keyId) {
+  return dispatch =>
+    apiClient.delete(`/account/api-keys/${keyId}`)
+      .then((response) => {
+        dispatch({
+          type: ActionTypes.API_KEY_REMOVED,
+          user: response.data
+        });
+      })
+      .catch((error) => {
+        const { response } = error;
+        Promise.reject(new Error(response.data.error));
+      });
+}
+
+export function unlinkService(service) {
+  return (dispatch) => {
+    if (!['github', 'google'].includes(service)) return;
+    apiClient.delete(`/auth/${service}`)
+      .then((response) => {
+        dispatch({
+          type: ActionTypes.AUTH_USER,
+          user: response.data
+        });
+      }).catch((error) => {
+        const { response } = error;
+        const message = response.message || response.data.error;
+        dispatch(authError(message));
+      });
   };
 }
