@@ -1,48 +1,75 @@
-import React, { PropTypes } from 'react';
+import PropTypes from 'prop-types';
+import React from 'react';
+import { withTranslation } from 'react-i18next';
+import Button from '../../../common/Button';
+
 import { domOnlyProps } from '../../../utils/reduxFormUtils';
 
 function LoginForm(props) {
-  const { fields: { email, password }, handleSubmit, submitting, pristine } = props;
+  const {
+    fields: { email, password },
+    handleSubmit,
+    submitting,
+    pristine,
+  } = props;
   return (
-    <form className="form" onSubmit={handleSubmit(props.validateAndLoginUser.bind(this, props.previousPath))}>
+    <form
+      className="form"
+      onSubmit={handleSubmit(props.validateAndLoginUser.bind(this, props.previousPath))}
+    >
       <p className="form__field">
-        <label htmlFor="email" className="form__label">Email or Username</label>
+        <label htmlFor="email" className="form__label">{props.t('LoginForm.UsernameOrEmail')}</label>
         <input
           className="form__input"
-          aria-label="email or username"
+          aria-label={props.t('LoginForm.UsernameOrEmailARIA')}
           type="text"
           id="email"
           {...domOnlyProps(email)}
         />
-        {email.touched && email.error && <span className="form-error">{email.error}</span>}
+        {email.touched && email.error && (
+          <span className="form-error">{email.error}</span>
+        )}
       </p>
       <p className="form__field">
-        <label htmlFor="password" className="form__label">Password</label>
+        <label htmlFor="password" className="form__label">{props.t('LoginForm.Password')}</label>
         <input
           className="form__input"
-          aria-label="password"
+          aria-label={props.t('LoginForm.PasswordARIA')}
           type="password"
           id="password"
           {...domOnlyProps(password)}
         />
-        {password.touched && password.error && <span className="form-error">{password.error}</span>}
+        {password.touched && password.error && (
+          <span className="form-error">{password.error}</span>
+        )}
       </p>
-      <input type="submit" disabled={submitting || pristine} value="Log In" aria-label="login" />
+      <Button
+        type="submit"
+        disabled={submitting || pristine}
+      >{props.t('LoginForm.Submit')}
+      </Button>
     </form>
   );
 }
 
 LoginForm.propTypes = {
   fields: PropTypes.shape({
-    email: PropTypes.object.isRequired,
-    password: PropTypes.object.isRequired
+    email: PropTypes.object.isRequired, // eslint-disable-line
+    password: PropTypes.object.isRequired, // eslint-disable-line
   }).isRequired,
   handleSubmit: PropTypes.func.isRequired,
   validateAndLoginUser: PropTypes.func.isRequired,
   submitting: PropTypes.bool,
-  invalid: PropTypes.bool,
   pristine: PropTypes.bool,
-  previousPath: PropTypes.string.isRequired
+  invalid: PropTypes.bool,
+  previousPath: PropTypes.string.isRequired,
+  t: PropTypes.func.isRequired
 };
 
-export default LoginForm;
+LoginForm.defaultProps = {
+  submitting: false,
+  pristine: true,
+  invalid: false,
+};
+
+export default withTranslation()(LoginForm);
