@@ -1,21 +1,36 @@
-import React, { PropTypes } from 'react';
+import PropTypes from 'prop-types';
+import React from 'react';
+import { withTranslation } from 'react-i18next';
 import { domOnlyProps } from '../../../utils/reduxFormUtils';
+import Button from '../../../common/Button';
 
 function ResetPasswordForm(props) {
-  const { fields: { email }, handleSubmit, submitting, invalid, pristine } = props;
+  const {
+    fields: { email }, handleSubmit, submitting, invalid, pristine, t
+  } = props;
   return (
-    <form className="form" onSubmit={handleSubmit(props.initiateResetPassword.bind(this))}>
+    <form
+      className="form"
+      onSubmit={handleSubmit(props.initiateResetPassword.bind(this))}
+    >
       <p className="form__field">
-        <label htmlFor="email" className="form__label">Email used for registration</label>
+        <label htmlFor="email" className="form__label">{t('ResetPasswordForm.Email')}</label>
         <input
           className="form__input"
-          aria-label="email"
+          aria-label={t('ResetPasswordForm.EmailARIA')}
           type="text"
           id="email"
           {...domOnlyProps(email)}
         />
+        {email.touched && email.error && (
+          <span className="form-error">{email.error}</span>
+        )}
       </p>
-      <input type="submit" disabled={submitting || invalid || pristine || props.user.resetPasswordInitiate} value="Send Password Reset Email" aria-label="Send email to reset password" />
+      <Button
+        type="submit"
+        disabled={submitting || invalid || pristine || props.user.resetPasswordInitiate}
+      >{t('ResetPasswordForm.Submit')}
+      </Button>
     </form>
   );
 }
@@ -31,7 +46,14 @@ ResetPasswordForm.propTypes = {
   pristine: PropTypes.bool,
   user: PropTypes.shape({
     resetPasswordInitiate: PropTypes.bool
-  })
+  }).isRequired,
+  t: PropTypes.func.isRequired
 };
 
-export default ResetPasswordForm;
+ResetPasswordForm.defaultProps = {
+  submitting: false,
+  pristine: true,
+  invalid: false,
+};
+
+export default withTranslation()(ResetPasswordForm);
