@@ -1,4 +1,5 @@
 import { combineReducers } from 'redux';
+import { reducer as form } from 'redux-form';
 import files from './modules/IDE/reducers/files';
 import ide from './modules/IDE/reducers/ide';
 import preferences from './modules/IDE/reducers/preferences';
@@ -8,7 +9,11 @@ import user from './modules/User/reducers';
 import sketches from './modules/IDE/reducers/projects';
 import toast from './modules/IDE/reducers/toast';
 import console from './modules/IDE/reducers/console';
-import { reducer as form } from 'redux-form';
+import assets from './modules/IDE/reducers/assets';
+import search from './modules/IDE/reducers/search';
+import sorting from './modules/IDE/reducers/sorting';
+import loading from './modules/IDE/reducers/loading';
+import collections from './modules/IDE/reducers/collections';
 
 const rootReducer = combineReducers({
   form,
@@ -18,9 +23,14 @@ const rootReducer = combineReducers({
   user,
   project,
   sketches,
+  search,
+  sorting,
   editorAccessibility,
   toast,
-  console
+  console,
+  assets,
+  loading,
+  collections
 });
 
 export default rootReducer;
