@@ -1,16 +1,22 @@
+import jsdom, { serializeDocument } from 'jsdom';
 import Project from '../models/project';
 import {
   injectMediaUrls,
   resolvePathsForElementsWithAttribute,
   resolveScripts,
-  resolveStyles } from '../utils/previewGeneration';
-import jsdom, { serializeDocument } from 'jsdom';
+  resolveStyles
+} from '../utils/previewGeneration';
+import { get404Sketch } from '../views/404Page';
 
 export function serveProject(req, res) {
   Project.findById(req.params.project_id)
     .exec((err, project) => {
+      if (err || !project) {
+        get404Sketch(html => res.send(html));
+        return;
+      }
       // TODO this does not parse html
-      const files = project.files;
+      const { files } = project;
       const htmlFile = files.find(file => file.name.match(/\.html$/i)).content;
       const filesToInject = files.filter(file => file.name.match(/\.(js|css)$/i));
       injectMediaUrls(filesToInject, files, req.params.project_id);
@@ -28,7 +34,10 @@ export function serveProject(req, res) {
         resolveScripts(sketchDoc, files);
         resolveStyles(sketchDoc, files);
 
+        res.setHeader('Cache-Control', 'public, max-age=0');
         res.send(serializeDocument(sketchDoc));
       });
     });
 }
+
+export default serveProject;
