@@ -1,46 +1,101 @@
-import User from '../models/user';
 import crypto from 'crypto';
 import async from 'async';
-import nodemailer from 'nodemailer';
-import mg from 'nodemailer-mailgun-transport';
+
+import User from '../models/user';
+import mail from '../utils/mail';
+import {
+  renderEmailConfirmation,
+  renderResetPassword,
+} from '../views/mail';
+
+export * from './user.controller/apiKey';
+
+export function userResponse(user) {
+  return {
+    email: user.email,
+    username: user.username,
+    preferences: user.preferences,
+    apiKeys: user.apiKeys,
+    verified: user.verified,
+    id: user._id,
+    totalSize: user.totalSize,
+    github: user.github,
+    google: user.google
+  };
+}
+
+const random = (done) => {
+  crypto.randomBytes(20, (err, buf) => {
+    const token = buf.toString('hex');
+    done(err, token);
+  });
+};
+
+export function findUserByUsername(username, cb) {
+  User.findByUsername(username, (err, user) => {
+    cb(user);
+  });
+}
 
 export function createUser(req, res, next) {
-  const user = new User({
-    username: req.body.username,
-    email: req.body.email,
-    password: req.body.password
-  });
+  const { username, email } = req.body;
+  const { password } = req.body;
+  const emailLowerCase = email.toLowerCase();
+  const EMAIL_VERIFY_TOKEN_EXPIRY_TIME = Date.now() + (3600000 * 24); // 24 hours
+  random((tokenError, token) => {
+    const user = new User({
+      username,
+      email: emailLowerCase,
+      password,
+      verified: User.EmailConfirmation.Sent,
+      verifiedToken: token,
+      verifiedTokenExpires: EMAIL_VERIFY_TOKEN_EXPIRY_TIME,
+    });
 
-  User.findOne({ email: req.body.email },
-    (err, existingUser) => { // eslint-disable-line consistent-return
-      if (err) { res.status(404).send({ error: err }); }
+    User.findByEmailAndUsername(email, username, (err, existingUser) => {
+      if (err) {
+        res.status(404).send({ error: err });
+        return;
+      }
 
       if (existingUser) {
-        return res.status(422).send({ error: 'Email is in use' });
+        const fieldInUse = existingUser.email.toLowerCase() === emailLowerCase ? 'Email' : 'Username';
+        res.status(422).send({ error: `${fieldInUse} is in use` });
+        return;
       }
-      user.save((saveErr) => { // eslint-disable-line consistent-return
-        if (saveErr) { return next(saveErr); }
-        req.logIn(user, (loginErr) => { // eslint-disable-line consistent-return
+      user.save((saveErr) => {
+        if (saveErr) {
+          next(saveErr);
+          return;
+        }
+        req.logIn(user, (loginErr) => {
           if (loginErr) {
-            return next(loginErr);
+            next(loginErr);
+            return;
           }
-          res.json({
-            email: req.user.email,
-            username: req.user.username,
-            preferences: req.user.preferences,
-            id: req.user._id
+
+          const protocol = process.env.NODE_ENV === 'production' ? 'https' : 'http';
+          const mailOptions = renderEmailConfirmation({
+            body: {
+              domain: `${protocol}://${req.headers.host}`,
+              link: `${protocol}://${req.headers.host}/verify?t=${token}`
+            },
+            to: req.user.email,
+          });
+
+          mail.send(mailOptions, (mailErr, result) => { // eslint-disable-line no-unused-vars
+            res.json(userResponse(req.user));
           });
         });
       });
     });
+  });
 }
 
 export function duplicateUserCheck(req, res) {
   const checkType = req.query.check_type;
   const value = req.query[checkType];
-  const query = {};
-  query[checkType] = value;
-  User.findOne(query, (err, user) => {
+  User.findByEmailOrUsername(value, { caseInsensitive: true }, (err, user) => {
     if (user) {
       return res.json({
         exists: true,
@@ -58,107 +113,155 @@ export function duplicateUserCheck(req, res) {
 export function updatePreferences(req, res) {
   User.findById(req.user.id, (err, user) => {
     if (err) {
-      return res.status(500).json({ error: err });
+      res.status(500).json({ error: err });
+      return;
     }
     if (!user) {
-      return res.status(404).json({ error: 'Document not found' });
+      res.status(404).json({ error: 'Document not found' });
+      return;
     }
 
     const preferences = Object.assign({}, user.preferences, req.body.preferences);
     user.preferences = preferences;
 
-    user.save((err) => {
-      if (err) {
-        return res.status(500).json({ error: err });
+    user.save((saveErr) => {
+      if (saveErr) {
+        res.status(500).json({ error: saveErr });
+        return;
       }
 
-      return res.json(user.preferences);
+      res.json(user.preferences);
     });
   });
 }
 
 export function resetPasswordInitiate(req, res) {
   async.waterfall([
-    (done) => {
-      crypto.randomBytes(20, (err, buf) => {
-        const token = buf.toString('hex');
-        done(err, token);
-      });
-    },
+    random,
     (token, done) => {
-      User.findOne({ email: req.body.email }, (err, user) => {
+      User.findByEmail(req.body.email, (err, user) => {
         if (!user) {
-          return res.json({ success: true, message: 'If the email is registered with the editor, an email has been sent.' });
+          res.json({ success: true, message: 'If the email is registered with the editor, an email has been sent.' });
+          return;
         }
         user.resetPasswordToken = token;
         user.resetPasswordExpires = Date.now() + 3600000; // 1 hour
 
-        user.save((err) => {
-          done(err, token, user);
+        user.save((saveErr) => {
+          done(saveErr, token, user);
         });
       });
     },
     (token, user, done) => {
-      const auth = {
-        auth: {
-          api_key: process.env.MAILGUN_KEY,
-          domain: process.env.MAILGUN_DOMAIN
-        }
-      };
-
-      const transporter = nodemailer.createTransport(mg(auth));
-      const message = {
+      const protocol = process.env.NODE_ENV === 'production' ? 'https' : 'http';
+      const mailOptions = renderResetPassword({
+        body: {
+          domain: `${protocol}://${req.headers.host}`,
+          link: `${protocol}://${req.headers.host}/reset-password/${token}`,
+        },
         to: user.email,
-        from: 'p5.js Web Editor <noreply@p5js.org>',
-        subject: 'p5.js Web Editor Password Reset',
-        text: `You are receiving this email because you (or someone else) have requested the reset of the password for your account.
-        \n\nPlease click on the following link, or paste this into your browser to complete the process:
-        \n\nhttp://${req.headers.host}/reset-password/${token}
-        \n\nIf you did not request this, please ignore this email and your password will remain unchanged.
-        \n\nThanks for using the p5.js Web Editor!\n`
-      };
-      transporter.sendMail(message, (error) => {
-        done(error);
       });
+
+      mail.send(mailOptions, done);
     }
   ], (err) => {
     if (err) {
       console.log(err);
-      return res.json({ success: false });
+      res.json({ success: false });
+      return;
     }
-    // send email here
-    return res.json({ success: true, message: 'If the email is registered with the editor, an email has been sent.' });
+    res.json({ success: true, message: 'If the email is registered with the editor, an email has been sent.' });
   });
 }
 
 export function validateResetPasswordToken(req, res) {
   User.findOne({ resetPasswordToken: req.params.token, resetPasswordExpires: { $gt: Date.now() } }, (err, user) => {
     if (!user) {
-      return res.status(401).json({ success: false, message: 'Password reset token is invalid or has expired.' });
+      res.status(401).json({ success: false, message: 'Password reset token is invalid or has expired.' });
+      return;
     }
     res.json({ success: true });
   });
 }
 
-export function updatePassword(req, res) {
-  User.findOne({ resetPasswordToken: req.params.token, resetPasswordExpires: { $gt: Date.now() } }, function (err, user) {
+export function emailVerificationInitiate(req, res) {
+  async.waterfall([
+    random,
+    (token, done) => {
+      User.findById(req.user.id, (err, user) => {
+        if (err) {
+          res.status(500).json({ error: err });
+          return;
+        }
+        if (!user) {
+          res.status(404).json({ error: 'Document not found' });
+          return;
+        }
+
+        if (user.verified === User.EmailConfirmation.Verified) {
+          res.status(409).json({ error: 'Email already verified' });
+          return;
+        }
+
+        const protocol = process.env.NODE_ENV === 'production' ? 'https' : 'http';
+        const mailOptions = renderEmailConfirmation({
+          body: {
+            domain: `${protocol}://${req.headers.host}`,
+            link: `${protocol}://${req.headers.host}/verify?t=${token}`
+          },
+          to: user.email,
+        });
+
+        mail.send(mailOptions, (mailErr, result) => { // eslint-disable-line no-unused-vars
+          if (mailErr != null) {
+            res.status(500).send({ error: 'Error sending mail' });
+          } else {
+            const EMAIL_VERIFY_TOKEN_EXPIRY_TIME = Date.now() + (3600000 * 24); // 24 hours
+            user.verified = User.EmailConfirmation.Resent;
+            user.verifiedToken = token;
+            user.verifiedTokenExpires = EMAIL_VERIFY_TOKEN_EXPIRY_TIME; // 24 hours
+            user.save();
+
+            res.json(userResponse(req.user));
+          }
+        });
+      });
+    },
+  ]);
+}
+
+export function verifyEmail(req, res) {
+  const token = req.query.t;
+
+  User.findOne({ verifiedToken: token, verifiedTokenExpires: { $gt: new Date() } }, (err, user) => {
     if (!user) {
-      return res.status(401).json({ success: false, message: 'Password reset token is invalid or has expired.' });
+      res.status(401).json({ success: false, message: 'Token is invalid or has expired.' });
+      return;
+    }
+
+    user.verified = User.EmailConfirmation.Verified;
+    user.verifiedToken = null;
+    user.verifiedTokenExpires = null;
+    user.save()
+      .then((result) => { // eslint-disable-line
+        res.json({ success: true });
+      });
+  });
+}
+
+export function updatePassword(req, res) {
+  User.findOne({ resetPasswordToken: req.params.token, resetPasswordExpires: { $gt: Date.now() } }, (err, user) => {
+    if (!user) {
+      res.status(401).json({ success: false, message: 'Password reset token is invalid or has expired.' });
+      return;
     }
 
     user.password = req.body.password;
     user.resetPasswordToken = undefined;
     user.resetPasswordExpires = undefined;
 
-    user.save(function (err) {
-      req.logIn(user, function (err) {
-        return res.json({
-          email: req.user.email,
-          username: req.user.username,
-          preferences: req.user.preferences,
-          id: req.user._id
-        });
-      });
+    user.save((saveErr) => {
+      req.logIn(user, loginErr => res.json(userResponse(req.user)));
     });
   });
 
@@ -166,7 +269,91 @@ export function updatePassword(req, res) {
 }
 
 export function userExists(username, callback) {
-  User.findOne({ username }, (err, user) => (
+  User.findByUsername(username, (err, user) => (
     user ? callback(true) : callback(false)
   ));
 }
+
+export function saveUser(res, user) {
+  user.save((saveErr) => {
+    if (saveErr) {
+      res.status(500).json({ error: saveErr });
+      return;
+    }
+
+    res.json(userResponse(user));
+  });
+}
+
+export function updateSettings(req, res) {
+  User.findById(req.user.id, (err, user) => {
+    if (err) {
+      res.status(500).json({ error: err });
+      return;
+    }
+    if (!user) {
+      res.status(404).json({ error: 'Document not found' });
+      return;
+    }
+
+    user.username = req.body.username;
+
+    if (req.body.currentPassword) {
+      user.comparePassword(req.body.currentPassword, (passwordErr, isMatch) => {
+        if (passwordErr) throw passwordErr;
+        if (!isMatch) {
+          res.status(401).json({ error: 'Current password is invalid.' });
+          return;
+        }
+        user.password = req.body.newPassword;
+        saveUser(res, user);
+      });
+    } else if (user.email !== req.body.email) {
+      const EMAIL_VERIFY_TOKEN_EXPIRY_TIME = Date.now() + (3600000 * 24); // 24 hours
+      user.verified = User.EmailConfirmation.Sent;
+
+      user.email = req.body.email;
+
+      random((error, token) => {
+        user.verifiedToken = token;
+        user.verifiedTokenExpires = EMAIL_VERIFY_TOKEN_EXPIRY_TIME;
+
+        saveUser(res, user);
+
+        const protocol = process.env.NODE_ENV === 'production' ? 'https' : 'http';
+        const mailOptions = renderEmailConfirmation({
+          body: {
+            domain: `${protocol}://${req.headers.host}`,
+            link: `${protocol}://${req.headers.host}/verify?t=${token}`
+          },
+          to: user.email,
+        });
+
+        mail.send(mailOptions);
+      });
+    } else {
+      saveUser(res, user);
+    }
+  });
+}
+
+export function unlinkGithub(req, res) {
+  if (req.user) {
+    req.user.github = undefined;
+    req.user.tokens = req.user.tokens.filter(token => token.kind !== 'github');
+    saveUser(res, req.user);
+    return;
+  }
+  res.status(404).json({ success: false, message: 'You must be logged in to complete this action.' });
+}
+
+export function unlinkGoogle(req, res) {
+  if (req.user) {
+    req.user.google = undefined;
+    req.user.tokens = req.user.tokens.filter(token => token.kind !== 'google');
+    saveUser(res, req.user);
+    return;
+  }
+  res.status(404).json({ success: false, message: 'You must be logged in to complete this action.' });
+}
+
