@@ -3,7 +3,8 @@ import * as EmbedController from '../controllers/embed.controller';
 
 const router = new Router();
 
-router.route('/embed/:project_id').get(EmbedController.serveProject);
-router.route('/full/:project_id').get(EmbedController.serveProject);
+router.get('/:username/embed/:project_id', EmbedController.serveProject);
+router.get('/:username/present/:project_id', EmbedController.serveProject);
+router.get('/embed/:project_id', EmbedController.serveProject);
 
 export default router;
