@@ -1,10 +1,10 @@
 import { Router } from 'express';
 import * as FileController from '../controllers/file.controller';
+import isAuthenticated from '../utils/isAuthenticated';
 
 const router = new Router();
 
-router.route('/projects/:project_id/files').post(FileController.createFile);
-router.route('/projects/:project_id/files/:file_id').delete(FileController.deleteFile);
-router.route('/projects/:project_id/*?').get(FileController.getFileContent);
+router.post('/projects/:project_id/files', isAuthenticated, FileController.createFile);
+router.delete('/projects/:project_id/files/:file_id', isAuthenticated, FileController.deleteFile);
 
 export default router;
